+++ Cargo.toml
@@ -25,3 +25,5 @@ tower = { version = "0.5.2", features = ["limit"] }
 tracing = "0.1.41"
 tracing-subscriber = "0.3.19"
 uuid = { version = "1.16.0", features = ["v4", "fast-rng"] }
+cfg-if = "1.0"
+
