+++ gateway-api/Cargo.toml
@@ -23,6 +23,7 @@ schemars.workspace = true
 serde_json.workspace = true
 serde.workspace = true
 serde_yaml.workspace = true
+cfg-if.workspace = true
 
 [dev-dependencies]
 k8s-openapi = { workspace = true, features = [ "v1_32", "schemars" ] }
@@ -36,3 +37,7 @@ uuid.workspace = true
 
 [package.metadata.docs.rs]
 features = [ "k8s-openapi/v1_32" ]
+
+[features]
+default = []
+experimental = []
