--- gateway-api/src/apis/mod.rs
@@ -1,2 +1,0 @@
-pub mod experimental;
-pub mod standard;
