+++ gateway-api/src/experimental/gatewayclasses.rs
@@ -1,6 +1,6 @@
 // WARNING: generated by kopium - manual changes will be overwritten
 // kopium command: kopium --schema=derived --derive=JsonSchema --derive=Default --derive=PartialEq --docs -f -
-// kopium version: 0.21.0
+// kopium version: 0.21.1
 
 #[allow(unused_imports)]
 mod prelude {
@@ -108,4 +108,20 @@ pub struct GatewayClassStatus {
     /// of GatewayClassConditionType for the type of each Condition.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub conditions: Option<Vec<Condition>>,
+    /// SupportedFeatures is the set of features the GatewayClass support.
+    /// It MUST be sorted in ascending alphabetical order by the Name key.
+    ///
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "supportedFeatures"
+    )]
+    pub supported_features: Option<Vec<GatewayClassStatusSupportedFeatures>>,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
+pub struct GatewayClassStatusSupportedFeatures {
+    /// FeatureName is used to describe distinct features that are covered by
+    /// conformance tests.
+    pub name: String,
 }
