+++ gateway-api/src/experimental/gateways.rs
@@ -1,6 +1,6 @@
 // WARNING: generated by kopium - manual changes will be overwritten
 // kopium command: kopium --schema=derived --derive=JsonSchema --derive=Default --derive=PartialEq --docs -f -
-// kopium version: 0.21.0
+// kopium version: 0.21.1
 
 #[allow(unused_imports)]
 mod prelude {
@@ -49,6 +49,18 @@ pub struct GatewaySpec {
     ///
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub addresses: Option<Vec<GatewayAddresses>>,
+    /// BackendTLS configures TLS settings for when this Gateway is connecting to
+    /// backends with TLS.
+    ///
+    /// Support: Core
+    ///
+    ///
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "backendTLS"
+    )]
+    pub backend_tls: Option<GatewayBackendTls>,
     /// GatewayClassName used for this Gateway. This is the name of a
     /// GatewayClass resource.
     #[serde(rename = "gatewayClassName")]
@@ -189,6 +201,80 @@ pub struct GatewayAddresses {
     ///
     /// Examples: `1.2.3.4`, `128::1`, `my-ip-address`.
     pub value: String,
+}
+
+/// BackendTLS configures TLS settings for when this Gateway is connecting to
+/// backends with TLS.
+///
+/// Support: Core
+///
+///
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
+pub struct GatewayBackendTls {
+    /// ClientCertificateRef is a reference to an object that contains a Client
+    /// Certificate and the associated private key.
+    ///
+    /// References to a resource in different namespace are invalid UNLESS there
+    /// is a ReferenceGrant in the target namespace that allows the certificate
+    /// to be attached. If a ReferenceGrant does not allow this reference, the
+    /// "ResolvedRefs" condition MUST be set to False for this listener with the
+    /// "RefNotPermitted" reason.
+    ///
+    /// ClientCertificateRef can reference to standard Kubernetes resources, i.e.
+    /// Secret, or implementation-specific custom resources.
+    ///
+    /// This setting can be overridden on the service level by use of BackendTLSPolicy.
+    ///
+    /// Support: Core
+    ///
+    ///
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "clientCertificateRef"
+    )]
+    pub client_certificate_ref: Option<GatewayBackendTlsClientCertificateRef>,
+}
+
+/// ClientCertificateRef is a reference to an object that contains a Client
+/// Certificate and the associated private key.
+///
+/// References to a resource in different namespace are invalid UNLESS there
+/// is a ReferenceGrant in the target namespace that allows the certificate
+/// to be attached. If a ReferenceGrant does not allow this reference, the
+/// "ResolvedRefs" condition MUST be set to False for this listener with the
+/// "RefNotPermitted" reason.
+///
+/// ClientCertificateRef can reference to standard Kubernetes resources, i.e.
+/// Secret, or implementation-specific custom resources.
+///
+/// This setting can be overridden on the service level by use of BackendTLSPolicy.
+///
+/// Support: Core
+///
+///
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
+pub struct GatewayBackendTlsClientCertificateRef {
+    /// Group is the group of the referent. For example, "gateway.networking.k8s.io".
+    /// When unspecified or empty string, core API group is inferred.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub group: Option<String>,
+    /// Kind is kind of the referent. For example "Secret".
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub kind: Option<String>,
+    /// Name is the name of the referent.
+    pub name: String,
+    /// Namespace is the namespace of the referenced object. When unspecified, the local
+    /// namespace is inferred.
+    ///
+    /// Note that when a namespace different than the local namespace is specified,
+    /// a ReferenceGrant object is required in the referent namespace to allow that
+    /// namespace's owner to accept the reference. See the ReferenceGrant
+    /// documentation for details.
+    ///
+    /// Support: Core
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub namespace: Option<String>,
 }
 
 /// Infrastructure defines infrastructure level attributes about this Gateway instance.
@@ -530,6 +616,21 @@ pub struct GatewayListenersTls {
         rename = "certificateRefs"
     )]
     pub certificate_refs: Option<Vec<GatewayListenersTlsCertificateRefs>>,
+    /// FrontendValidation holds configuration information for validating the frontend (client).
+    /// Setting this field will require clients to send a client certificate
+    /// required for validation during the TLS handshake. In browsers this may result in a dialog appearing
+    /// that requests a user to specify the client certificate.
+    /// The maximum depth of a certificate chain accepted in verification is Implementation specific.
+    ///
+    /// Support: Extended
+    ///
+    ///
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "frontendValidation"
+    )]
+    pub frontend_validation: Option<GatewayListenersTlsFrontendValidation>,
     /// Mode defines the TLS behavior for the TLS session initiated by the client.
     /// There are two possible modes:
     ///
@@ -577,6 +678,76 @@ pub struct GatewayListenersTlsCertificateRefs {
     /// Kind is kind of the referent. For example "Secret".
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub kind: Option<String>,
+    /// Name is the name of the referent.
+    pub name: String,
+    /// Namespace is the namespace of the referenced object. When unspecified, the local
+    /// namespace is inferred.
+    ///
+    /// Note that when a namespace different than the local namespace is specified,
+    /// a ReferenceGrant object is required in the referent namespace to allow that
+    /// namespace's owner to accept the reference. See the ReferenceGrant
+    /// documentation for details.
+    ///
+    /// Support: Core
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub namespace: Option<String>,
+}
+
+/// FrontendValidation holds configuration information for validating the frontend (client).
+/// Setting this field will require clients to send a client certificate
+/// required for validation during the TLS handshake. In browsers this may result in a dialog appearing
+/// that requests a user to specify the client certificate.
+/// The maximum depth of a certificate chain accepted in verification is Implementation specific.
+///
+/// Support: Extended
+///
+///
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
+pub struct GatewayListenersTlsFrontendValidation {
+    /// CACertificateRefs contains one or more references to
+    /// Kubernetes objects that contain TLS certificates of
+    /// the Certificate Authorities that can be used
+    /// as a trust anchor to validate the certificates presented by the client.
+    ///
+    /// A single CA certificate reference to a Kubernetes ConfigMap
+    /// has "Core" support.
+    /// Implementations MAY choose to support attaching multiple CA certificates to
+    /// a Listener, but this behavior is implementation-specific.
+    ///
+    /// Support: Core - A single reference to a Kubernetes ConfigMap
+    /// with the CA certificate in a key named `ca.crt`.
+    ///
+    /// Support: Implementation-specific (More than one reference, or other kinds
+    /// of resources).
+    ///
+    /// References to a resource in a different namespace are invalid UNLESS there
+    /// is a ReferenceGrant in the target namespace that allows the certificate
+    /// to be attached. If a ReferenceGrant does not allow this reference, the
+    /// "ResolvedRefs" condition MUST be set to False for this listener with the
+    /// "RefNotPermitted" reason.
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "caCertificateRefs"
+    )]
+    pub ca_certificate_refs: Option<Vec<GatewayListenersTlsFrontendValidationCaCertificateRefs>>,
+}
+
+/// ObjectReference identifies an API object including its namespace.
+///
+/// The API object must be valid in the cluster; the Group and Kind must
+/// be registered in the cluster for this reference to be valid.
+///
+/// References to objects with invalid Group and Kind are not valid, and must
+/// be rejected by the implementation, with appropriate Conditions set
+/// on the containing object.
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
+pub struct GatewayListenersTlsFrontendValidationCaCertificateRefs {
+    /// Group is the group of the referent. For example, "gateway.networking.k8s.io".
+    /// When unspecified or empty string, core API group is inferred.
+    pub group: String,
+    /// Kind is kind of the referent. For example "ConfigMap" or "Service".
+    pub kind: String,
     /// Name is the name of the referent.
     pub name: String,
     /// Namespace is the namespace of the referenced object. When unspecified, the local
