+++ gateway-api/src/experimental/httproutes.rs
@@ -1,6 +1,6 @@
 // WARNING: generated by kopium - manual changes will be overwritten
 // kopium command: kopium --schema=derived --derive=JsonSchema --derive=Default --derive=PartialEq --docs -f -
-// kopium version: 0.21.0
+// kopium version: 0.21.1
 
 #[allow(unused_imports)]
 mod prelude {
@@ -131,6 +131,16 @@ pub struct HTTPRouteSpec {
     /// generic way to enable other kinds of cross-namespace reference.
     ///
     ///
+    /// ParentRefs from a Route to a Service in the same namespace are "producer"
+    /// routes, which apply default routing rules to inbound connections from
+    /// any namespace to the Service.
+    ///
+    /// ParentRefs from a Route to a Service in a different namespace are
+    /// "consumer" routes, and these routing rules are only applied to outbound
+    /// connections originating from the same namespace as the Route, for which
+    /// the intended destination of the connections are a Service targeted as a
+    /// ParentRef of the Route.
+    ///
     ///
     ///
     ///
@@ -195,6 +205,16 @@ pub struct HTTPRouteParentRefs {
     /// generic way to enable any other kind of cross-namespace reference.
     ///
     ///
+    /// ParentRefs from a Route to a Service in the same namespace are "producer"
+    /// routes, which apply default routing rules to inbound connections from
+    /// any namespace to the Service.
+    ///
+    /// ParentRefs from a Route to a Service in a different namespace are
+    /// "consumer" routes, and these routing rules are only applied to outbound
+    /// connections originating from the same namespace as the Route, for which
+    /// the intended destination of the connections are a Service targeted as a
+    /// ParentRef of the Route.
+    ///
     ///
     /// Support: Core
     #[serde(default, skip_serializing_if = "Option::is_none")]
@@ -210,6 +230,10 @@ pub struct HTTPRouteParentRefs {
     /// and SectionName are specified, the name and port of the selected listener
     /// must match both specified values.
     ///
+    ///
+    /// When the parent resource is a Service, this targets a specific port in the
+    /// Service spec. When both Port (experimental) and SectionName are specified,
+    /// the name and port of the selected port must match both specified values.
     ///
     ///
     /// Implementations MAY choose to support other parent resources.
@@ -349,7 +373,7 @@ pub struct HTTPRouteRules {
     ///
     /// For example, take the following matches configuration:
     ///
-    /// ```
+    /// ```text
     /// matches:
     /// - path:
     ///     value: "/foo"
@@ -401,6 +425,31 @@ pub struct HTTPRouteRules {
     /// parent a request is coming from, a HTTP 404 status code MUST be returned.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub matches: Option<Vec<HTTPRouteRulesMatches>>,
+    /// Name is the name of the route rule. This name MUST be unique within a Route if it is set.
+    ///
+    /// Support: Extended
+    ///
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub name: Option<String>,
+    /// Retry defines the configuration for when to retry an HTTP request.
+    ///
+    /// Support: Extended
+    ///
+    ///
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub retry: Option<HTTPRouteRulesRetry>,
+    /// SessionPersistence defines and configures session persistence
+    /// for the route rule.
+    ///
+    /// Support: Extended
+    ///
+    ///
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "sessionPersistence"
+    )]
+    pub session_persistence: Option<HTTPRouteRulesSessionPersistence>,
     /// Timeouts defines the timeouts that can be configured for an HTTP request.
     ///
     /// Support: Extended
@@ -763,6 +812,25 @@ pub struct HTTPRouteRulesBackendRefsFiltersRequestMirror {
     /// Support: Implementation-specific for any other resource
     #[serde(rename = "backendRef")]
     pub backend_ref: HTTPRouteRulesBackendRefsFiltersRequestMirrorBackendRef,
+    /// Fraction represents the fraction of requests that should be
+    /// mirrored to BackendRef.
+    ///
+    /// Only one of Fraction or Percent may be specified. If neither field
+    /// is specified, 100% of requests will be mirrored.
+    ///
+    ///
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub fraction: Option<HTTPRouteRulesBackendRefsFiltersRequestMirrorFraction>,
+    /// Percent represents the percentage of requests that should be
+    /// mirrored to BackendRef. Its minimum value is 0 (indicating 0% of
+    /// requests) and its maximum value is 100 (indicating 100% of requests).
+    ///
+    /// Only one of Fraction or Percent may be specified. If neither field
+    /// is specified, 100% of requests will be mirrored.
+    ///
+    ///
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub percent: Option<i32>,
 }
 
 /// BackendRef references a resource where mirrored requests are sent.
@@ -830,6 +898,20 @@ pub struct HTTPRouteRulesBackendRefsFiltersRequestMirrorBackendRef {
     /// resource or this field.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub port: Option<i32>,
+}
+
+/// Fraction represents the fraction of requests that should be
+/// mirrored to BackendRef.
+///
+/// Only one of Fraction or Percent may be specified. If neither field
+/// is specified, 100% of requests will be mirrored.
+///
+///
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
+pub struct HTTPRouteRulesBackendRefsFiltersRequestMirrorFraction {
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub denominator: Option<i32>,
+    pub numerator: i32,
 }
 
 /// RequestRedirect defines a schema for a filter that responds to the
@@ -1446,6 +1528,25 @@ pub struct HTTPRouteRulesFiltersRequestMirror {
     /// Support: Implementation-specific for any other resource
     #[serde(rename = "backendRef")]
     pub backend_ref: HTTPRouteRulesFiltersRequestMirrorBackendRef,
+    /// Fraction represents the fraction of requests that should be
+    /// mirrored to BackendRef.
+    ///
+    /// Only one of Fraction or Percent may be specified. If neither field
+    /// is specified, 100% of requests will be mirrored.
+    ///
+    ///
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub fraction: Option<HTTPRouteRulesFiltersRequestMirrorFraction>,
+    /// Percent represents the percentage of requests that should be
+    /// mirrored to BackendRef. Its minimum value is 0 (indicating 0% of
+    /// requests) and its maximum value is 100 (indicating 100% of requests).
+    ///
+    /// Only one of Fraction or Percent may be specified. If neither field
+    /// is specified, 100% of requests will be mirrored.
+    ///
+    ///
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub percent: Option<i32>,
 }
 
 /// BackendRef references a resource where mirrored requests are sent.
@@ -1513,6 +1614,20 @@ pub struct HTTPRouteRulesFiltersRequestMirrorBackendRef {
     /// resource or this field.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub port: Option<i32>,
+}
+
+/// Fraction represents the fraction of requests that should be
+/// mirrored to BackendRef.
+///
+/// Only one of Fraction or Percent may be specified. If neither field
+/// is specified, 100% of requests will be mirrored.
+///
+///
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
+pub struct HTTPRouteRulesFiltersRequestMirrorFraction {
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub denominator: Option<i32>,
+    pub numerator: i32,
 }
 
 /// RequestRedirect defines a schema for a filter that responds to the
@@ -1871,7 +1986,7 @@ pub enum HTTPRouteRulesFiltersUrlRewritePathType {
 /// For example, the match below will match a HTTP request only if its path
 /// starts with `/foo` AND it contains the `version: v1` header:
 ///
-/// ```
+/// ```text
 /// match:
 ///
 /// 	path:
@@ -1962,7 +2077,7 @@ pub enum HTTPRouteRulesMatchesHeadersType {
 /// For example, the match below will match a HTTP request only if its path
 /// starts with `/foo` AND it contains the `version: v1` header:
 ///
-/// ```
+/// ```text
 /// match:
 ///
 /// 	path:
@@ -2063,6 +2178,187 @@ pub struct HTTPRouteRulesMatchesQueryParams {
 pub enum HTTPRouteRulesMatchesQueryParamsType {
     Exact,
     RegularExpression,
+}
+
+/// Retry defines the configuration for when to retry an HTTP request.
+///
+/// Support: Extended
+///
+///
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
+pub struct HTTPRouteRulesRetry {
+    /// Attempts specifies the maxmimum number of times an individual request
+    /// from the gateway to a backend should be retried.
+    ///
+    /// If the maximum number of retries has been attempted without a successful
+    /// response from the backend, the Gateway MUST return an error.
+    ///
+    /// When this field is unspecified, the number of times to attempt to retry
+    /// a backend request is implementation-specific.
+    ///
+    /// Support: Extended
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub attempts: Option<i64>,
+    /// Backoff specifies the minimum duration a Gateway should wait between
+    /// retry attempts and is represented in Gateway API Duration formatting.
+    ///
+    /// For example, setting the `rules[].retry.backoff` field to the value
+    /// `100ms` will cause a backend request to first be retried approximately
+    /// 100 milliseconds after timing out or receiving a response code configured
+    /// to be retryable.
+    ///
+    /// An implementation MAY use an exponential or alternative backoff strategy
+    /// for subsequent retry attempts, MAY cap the maximum backoff duration to
+    /// some amount greater than the specified minimum, and MAY add arbitrary
+    /// jitter to stagger requests, as long as unsuccessful backend requests are
+    /// not retried before the configured minimum duration.
+    ///
+    /// If a Request timeout (`rules[].timeouts.request`) is configured on the
+    /// route, the entire duration of the initial request and any retry attempts
+    /// MUST not exceed the Request timeout duration. If any retry attempts are
+    /// still in progress when the Request timeout duration has been reached,
+    /// these SHOULD be canceled if possible and the Gateway MUST immediately
+    /// return a timeout error.
+    ///
+    /// If a BackendRequest timeout (`rules[].timeouts.backendRequest`) is
+    /// configured on the route, any retry attempts which reach the configured
+    /// BackendRequest timeout duration without a response SHOULD be canceled if
+    /// possible and the Gateway should wait for at least the specified backoff
+    /// duration before attempting to retry the backend request again.
+    ///
+    /// If a BackendRequest timeout is _not_ configured on the route, retry
+    /// attempts MAY time out after an implementation default duration, or MAY
+    /// remain pending until a configured Request timeout or implementation
+    /// default duration for total request time is reached.
+    ///
+    /// When this field is unspecified, the time to wait between retry attempts
+    /// is implementation-specific.
+    ///
+    /// Support: Extended
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub backoff: Option<String>,
+    /// Codes defines the HTTP response status codes for which a backend request
+    /// should be retried.
+    ///
+    /// Support: Extended
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub codes: Option<Vec<i64>>,
+}
+
+/// SessionPersistence defines and configures session persistence
+/// for the route rule.
+///
+/// Support: Extended
+///
+///
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
+pub struct HTTPRouteRulesSessionPersistence {
+    /// AbsoluteTimeout defines the absolute timeout of the persistent
+    /// session. Once the AbsoluteTimeout duration has elapsed, the
+    /// session becomes invalid.
+    ///
+    /// Support: Extended
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "absoluteTimeout"
+    )]
+    pub absolute_timeout: Option<String>,
+    /// CookieConfig provides configuration settings that are specific
+    /// to cookie-based session persistence.
+    ///
+    /// Support: Core
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "cookieConfig"
+    )]
+    pub cookie_config: Option<HTTPRouteRulesSessionPersistenceCookieConfig>,
+    /// IdleTimeout defines the idle timeout of the persistent session.
+    /// Once the session has been idle for more than the specified
+    /// IdleTimeout duration, the session becomes invalid.
+    ///
+    /// Support: Extended
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "idleTimeout"
+    )]
+    pub idle_timeout: Option<String>,
+    /// SessionName defines the name of the persistent session token
+    /// which may be reflected in the cookie or the header. Users
+    /// should avoid reusing session names to prevent unintended
+    /// consequences, such as rejection or unpredictable behavior.
+    ///
+    /// Support: Implementation-specific
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "sessionName"
+    )]
+    pub session_name: Option<String>,
+    /// Type defines the type of session persistence such as through
+    /// the use a header or cookie. Defaults to cookie based session
+    /// persistence.
+    ///
+    /// Support: Core for "Cookie" type
+    ///
+    /// Support: Extended for "Header" type
+    #[serde(default, skip_serializing_if = "Option::is_none", rename = "type")]
+    pub r#type: Option<HTTPRouteRulesSessionPersistenceType>,
+}
+
+/// CookieConfig provides configuration settings that are specific
+/// to cookie-based session persistence.
+///
+/// Support: Core
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
+pub struct HTTPRouteRulesSessionPersistenceCookieConfig {
+    /// LifetimeType specifies whether the cookie has a permanent or
+    /// session-based lifetime. A permanent cookie persists until its
+    /// specified expiry time, defined by the Expires or Max-Age cookie
+    /// attributes, while a session cookie is deleted when the current
+    /// session ends.
+    ///
+    /// When set to "Permanent", AbsoluteTimeout indicates the
+    /// cookie's lifetime via the Expires or Max-Age cookie attributes
+    /// and is required.
+    ///
+    /// When set to "Session", AbsoluteTimeout indicates the
+    /// absolute lifetime of the cookie tracked by the gateway and
+    /// is optional.
+    ///
+    /// Support: Core for "Session" type
+    ///
+    /// Support: Extended for "Permanent" type
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "lifetimeType"
+    )]
+    pub lifetime_type: Option<HTTPRouteRulesSessionPersistenceCookieConfigLifetimeType>,
+}
+
+/// CookieConfig provides configuration settings that are specific
+/// to cookie-based session persistence.
+///
+/// Support: Core
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, PartialEq)]
+pub enum HTTPRouteRulesSessionPersistenceCookieConfigLifetimeType {
+    Permanent,
+    Session,
+}
+
+/// SessionPersistence defines and configures session persistence
+/// for the route rule.
+///
+/// Support: Extended
+///
+///
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, PartialEq)]
+pub enum HTTPRouteRulesSessionPersistenceType {
+    Cookie,
+    Header,
 }
 
 /// Timeouts defines the timeouts that can be configured for an HTTP request.
@@ -2223,6 +2519,16 @@ pub struct HTTPRouteStatusParentsParentRef {
     /// generic way to enable any other kind of cross-namespace reference.
     ///
     ///
+    /// ParentRefs from a Route to a Service in the same namespace are "producer"
+    /// routes, which apply default routing rules to inbound connections from
+    /// any namespace to the Service.
+    ///
+    /// ParentRefs from a Route to a Service in a different namespace are
+    /// "consumer" routes, and these routing rules are only applied to outbound
+    /// connections originating from the same namespace as the Route, for which
+    /// the intended destination of the connections are a Service targeted as a
+    /// ParentRef of the Route.
+    ///
     ///
     /// Support: Core
     #[serde(default, skip_serializing_if = "Option::is_none")]
@@ -2238,6 +2544,10 @@ pub struct HTTPRouteStatusParentsParentRef {
     /// and SectionName are specified, the name and port of the selected listener
     /// must match both specified values.
     ///
+    ///
+    /// When the parent resource is a Service, this targets a specific port in the
+    /// Service spec. When both Port (experimental) and SectionName are specified,
+    /// the name and port of the selected port must match both specified values.
     ///
     ///
     /// Implementations MAY choose to support other parent resources.
