+++ gateway-api/src/experimental/tlsroutes.rs
@@ -1,6 +1,6 @@
 // WARNING: generated by kopium - manual changes will be overwritten
 // kopium command: kopium --schema=derived --derive=JsonSchema --derive=Default --derive=PartialEq --docs -f -
-// kopium version: 0.21.0
+// kopium version: 0.21.1
 
 #[allow(unused_imports)]
 mod prelude {
