+++ gateway-api/src/lib.rs
@@ -1,7 +1,15 @@
-pub mod apis;
-
 pub mod duration;
 pub use duration::Duration;
+pub mod standard;
+pub use standard::*;
+
+cfg_if::cfg_if! {
+   if #[cfg(feature = "experimental")] {
+        pub mod experimental;
+    } else {
+
+    }
+}
 
 #[cfg(test)]
 mod tests {
@@ -22,7 +30,7 @@ mod tests {
     use tower::ServiceBuilder;
     use uuid::Uuid;
 
-    use crate::apis::standard::{
+    use crate::{
         constants::{
             GatewayConditionReason, GatewayConditionType, ListenerConditionReason,
             ListenerConditionType,
