+++ gateway-api/src/standard/constants.rs
