+++ gateway-api/src/standard/grpcroutes.rs
@@ -1,6 +1,6 @@
 // WARNING: generated by kopium - manual changes will be overwritten
 // kopium command: kopium --schema=derived --derive=JsonSchema --derive=Default --derive=PartialEq --docs -f -
-// kopium version: 0.21.0
+// kopium version: 0.21.1
 
 #[allow(unused_imports)]
 mod prelude {
@@ -327,7 +327,7 @@ pub struct GRPCRouteRules {
     ///
     /// For example, take the following matches configuration:
     ///
-    /// ```
+    /// ```text
     /// matches:
     /// - method:
     ///     service: foo.bar
@@ -1305,7 +1305,7 @@ pub enum GRPCRouteRulesFiltersType {
 /// For example, the match below will match a gRPC request only if its service
 /// is `foo` AND it contains the `version: v1` header:
 ///
-/// ```
+/// ```text
 /// matches:
 ///   - method:
 ///     type: Exact
