+++ gateway-api/src/standard/httproutes.rs
@@ -1,6 +1,6 @@
 // WARNING: generated by kopium - manual changes will be overwritten
 // kopium command: kopium --schema=derived --derive=JsonSchema --derive=Default --derive=PartialEq --docs -f -
-// kopium version: 0.21.0
+// kopium version: 0.21.1
 
 #[allow(unused_imports)]
 mod prelude {
@@ -11,67 +11,70 @@ mod prelude {
 }
 use self::prelude::*;
 
-/// Spec defines the desired state of GRPCRoute.
+/// Spec defines the desired state of HTTPRoute.
 #[derive(CustomResource, Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
 #[kube(
     group = "gateway.networking.k8s.io",
     version = "v1",
-    kind = "GRPCRoute",
-    plural = "grpcroutes"
+    kind = "HTTPRoute",
+    plural = "httproutes"
 )]
 #[kube(namespaced)]
-#[kube(status = "GRPCRouteStatus")]
+#[kube(status = "HTTPRouteStatus")]
 #[kube(derive = "Default")]
 #[kube(derive = "PartialEq")]
-pub struct GRPCRouteSpec {
-    /// Hostnames defines a set of hostnames to match against the GRPC
-    /// Host header to select a GRPCRoute to process the request. This matches
-    /// the RFC 1123 definition of a hostname with 2 notable exceptions:
+pub struct HTTPRouteSpec {
+    /// Hostnames defines a set of hostnames that should match against the HTTP Host
+    /// header to select a HTTPRoute used to process the request. Implementations
+    /// MUST ignore any port value specified in the HTTP Host header while
+    /// performing a match and (absent of any applicable header modification
+    /// configuration) MUST forward this header unmodified to the backend.
+    ///
+    /// Valid values for Hostnames are determined by RFC 1123 definition of a
+    /// hostname with 2 notable exceptions:
     ///
     /// 1. IPs are not allowed.
     /// 2. A hostname may be prefixed with a wildcard label (`*.`). The wildcard
-    ///    label MUST appear by itself as the first label.
+    ///    label must appear by itself as the first label.
     ///
-    /// If a hostname is specified by both the Listener and GRPCRoute, there
-    /// MUST be at least one intersecting hostname for the GRPCRoute to be
+    /// If a hostname is specified by both the Listener and HTTPRoute, there
+    /// must be at least one intersecting hostname for the HTTPRoute to be
     /// attached to the Listener. For example:
     ///
-    /// * A Listener with `test.example.com` as the hostname matches GRPCRoutes
+    /// * A Listener with `test.example.com` as the hostname matches HTTPRoutes
     ///   that have either not specified any hostnames, or have specified at
     ///   least one of `test.example.com` or `*.example.com`.
-    /// * A Listener with `*.example.com` as the hostname matches GRPCRoutes
+    /// * A Listener with `*.example.com` as the hostname matches HTTPRoutes
     ///   that have either not specified any hostnames or have specified at least
     ///   one hostname that matches the Listener hostname. For example,
-    ///   `test.example.com` and `*.example.com` would both match. On the other
-    ///   hand, `example.com` and `test.example.net` would not match.
+    ///   `*.example.com`, `test.example.com`, and `foo.test.example.com` would
+    ///   all match. On the other hand, `example.com` and `test.example.net` would
+    ///   not match.
     ///
     /// Hostnames that are prefixed with a wildcard label (`*.`) are interpreted
     /// as a suffix match. That means that a match for `*.example.com` would match
     /// both `test.example.com`, and `foo.test.example.com`, but not `example.com`.
     ///
-    /// If both the Listener and GRPCRoute have specified hostnames, any
-    /// GRPCRoute hostnames that do not match the Listener hostname MUST be
+    /// If both the Listener and HTTPRoute have specified hostnames, any
+    /// HTTPRoute hostnames that do not match the Listener hostname MUST be
     /// ignored. For example, if a Listener specified `*.example.com`, and the
-    /// GRPCRoute specified `test.example.com` and `test.example.net`,
-    /// `test.example.net` MUST NOT be considered for a match.
+    /// HTTPRoute specified `test.example.com` and `test.example.net`,
+    /// `test.example.net` must not be considered for a match.
     ///
-    /// If both the Listener and GRPCRoute have specified hostnames, and none
-    /// match with the criteria above, then the GRPCRoute MUST NOT be accepted by
-    /// the implementation. The implementation MUST raise an 'Accepted' Condition
-    /// with a status of `False` in the corresponding RouteParentStatus.
+    /// If both the Listener and HTTPRoute have specified hostnames, and none
+    /// match with the criteria above, then the HTTPRoute is not accepted. The
+    /// implementation must raise an 'Accepted' Condition with a status of
+    /// `False` in the corresponding RouteParentStatus.
     ///
-    /// If a Route (A) of type HTTPRoute or GRPCRoute is attached to a
-    /// Listener and that listener already has another Route (B) of the other
-    /// type attached and the intersection of the hostnames of A and B is
-    /// non-empty, then the implementation MUST accept exactly one of these two
-    /// routes, determined by the following criteria, in order:
+    /// In the event that multiple HTTPRoutes specify intersecting hostnames (e.g.
+    /// overlapping wildcard matching and exact matching hostnames), precedence must
+    /// be given to rules from the HTTPRoute with the largest number of:
     ///
-    /// * The oldest Route based on creation timestamp.
-    /// * The Route appearing first in alphabetical order by
-    ///   "{namespace}/{name}".
+    /// * Characters in a matching non-wildcard hostname.
+    /// * Characters in a matching hostname.
     ///
-    /// The rejected Route MUST raise an 'Accepted' condition with a status of
-    /// 'False' in the corresponding RouteParentStatus.
+    /// If ties exist across multiple Routes, the matching precedence rules for
+    /// HTTPRouteMatches takes over.
     ///
     /// Support: Core
     #[serde(default, skip_serializing_if = "Option::is_none")]
@@ -128,16 +131,6 @@ pub struct GRPCRouteSpec {
     /// generic way to enable other kinds of cross-namespace reference.
     ///
     ///
-    /// ParentRefs from a Route to a Service in the same namespace are "producer"
-    /// routes, which apply default routing rules to inbound connections from
-    /// any namespace to the Service.
-    ///
-    /// ParentRefs from a Route to a Service in a different namespace are
-    /// "consumer" routes, and these routing rules are only applied to outbound
-    /// connections originating from the same namespace as the Route, for which
-    /// the intended destination of the connections are a Service targeted as a
-    /// ParentRef of the Route.
-    ///
     ///
     ///
     ///
@@ -148,12 +141,12 @@ pub struct GRPCRouteSpec {
         skip_serializing_if = "Option::is_none",
         rename = "parentRefs"
     )]
-    pub parent_refs: Option<Vec<GRPCRouteParentRefs>>,
-    /// Rules are a list of GRPC matchers, filters and actions.
+    pub parent_refs: Option<Vec<HTTPRouteParentRefs>>,
+    /// Rules are a list of HTTP matchers, filters and actions.
     ///
     ///
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub rules: Option<Vec<GRPCRouteRules>>,
+    pub rules: Option<Vec<HTTPRouteRules>>,
 }
 
 /// ParentReference identifies an API object (usually a Gateway) that can be considered
@@ -169,7 +162,7 @@ pub struct GRPCRouteSpec {
 /// The API object must be valid in the cluster; the Group and Kind must
 /// be registered in the cluster for this reference to be valid.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteParentRefs {
+pub struct HTTPRouteParentRefs {
     /// Group is the group of the referent.
     /// When unspecified, "gateway.networking.k8s.io" is inferred.
     /// To set the core API group (such as for a "Service" kind referent),
@@ -202,16 +195,6 @@ pub struct GRPCRouteParentRefs {
     /// generic way to enable any other kind of cross-namespace reference.
     ///
     ///
-    /// ParentRefs from a Route to a Service in the same namespace are "producer"
-    /// routes, which apply default routing rules to inbound connections from
-    /// any namespace to the Service.
-    ///
-    /// ParentRefs from a Route to a Service in a different namespace are
-    /// "consumer" routes, and these routing rules are only applied to outbound
-    /// connections originating from the same namespace as the Route, for which
-    /// the intended destination of the connections are a Service targeted as a
-    /// ParentRef of the Route.
-    ///
     ///
     /// Support: Core
     #[serde(default, skip_serializing_if = "Option::is_none")]
@@ -227,10 +210,6 @@ pub struct GRPCRouteParentRefs {
     /// and SectionName are specified, the name and port of the selected listener
     /// must match both specified values.
     ///
-    ///
-    /// When the parent resource is a Service, this targets a specific port in the
-    /// Service spec. When both Port (experimental) and SectionName are specified,
-    /// the name and port of the selected port must match both specified values.
     ///
     ///
     /// Implementations MAY choose to support other parent resources.
@@ -280,11 +259,11 @@ pub struct GRPCRouteParentRefs {
     pub section_name: Option<String>,
 }
 
-/// GRPCRouteRule defines the semantics for matching a gRPC request based on
+/// HTTPRouteRule defines semantics for matching an HTTP request based on
 /// conditions (matches), processing it (filters), and forwarding the request to
 /// an API object (backendRefs).
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRules {
+pub struct HTTPRouteRules {
     /// BackendRefs defines the backend(s) where matching requests should be
     /// sent.
     ///
@@ -293,22 +272,29 @@ pub struct GRPCRouteRules {
     ///
     /// If *all* entries in BackendRefs are invalid, and there are also no filters
     /// specified in this route rule, *all* traffic which matches this rule MUST
-    /// receive an `UNAVAILABLE` status.
+    /// receive a 500 status code.
     ///
-    /// See the GRPCBackendRef definition for the rules about what makes a single
-    /// GRPCBackendRef invalid.
+    /// See the HTTPBackendRef definition for the rules about what makes a single
+    /// HTTPBackendRef invalid.
     ///
-    /// When a GRPCBackendRef is invalid, `UNAVAILABLE` statuses MUST be returned for
+    /// When a HTTPBackendRef is invalid, 500 status codes MUST be returned for
     /// requests that would have otherwise been routed to an invalid backend. If
     /// multiple backends are specified, and some are invalid, the proportion of
     /// requests that would otherwise have been routed to an invalid backend
-    /// MUST receive an `UNAVAILABLE` status.
+    /// MUST receive a 500 status code.
     ///
     /// For example, if two backends are specified with equal weights, and one is
-    /// invalid, 50 percent of traffic MUST receive an `UNAVAILABLE` status.
-    /// Implementations may choose how that 50 percent is determined.
+    /// invalid, 50 percent of traffic must receive a 500. Implementations may
+    /// choose how that 50 percent is determined.
+    ///
+    /// When a HTTPBackendRef refers to a Service that has no ready endpoints,
+    /// implementations SHOULD return a 503 for requests to that backend instead.
+    /// If an implementation chooses to do this, all of the above rules for 500 responses
+    /// MUST also apply for responses that return a 503.
     ///
     /// Support: Core for Kubernetes Service
+    ///
+    /// Support: Extended for Kubernetes ServiceImport
     ///
     /// Support: Implementation-specific for any other resource
     ///
@@ -318,17 +304,27 @@ pub struct GRPCRouteRules {
         skip_serializing_if = "Option::is_none",
         rename = "backendRefs"
     )]
-    pub backend_refs: Option<Vec<GRPCRouteRulesBackendRefs>>,
+    pub backend_refs: Option<Vec<HTTPRouteRulesBackendRefs>>,
     /// Filters define the filters that are applied to requests that match
     /// this rule.
     ///
-    /// The effects of ordering of multiple behaviors are currently unspecified.
-    /// This can change in the future based on feedback during the alpha stage.
+    /// Wherever possible, implementations SHOULD implement filters in the order
+    /// they are specified.
+    ///
+    /// Implementations MAY choose to implement this ordering strictly, rejecting
+    /// any combination or order of filters that can not be supported. If implementations
+    /// choose a strict interpretation of filter ordering, they MUST clearly document
+    /// that behavior.
+    ///
+    /// To reject an invalid combination or order of filters, implementations SHOULD
+    /// consider the Route Rules with this configuration invalid. If all Route Rules
+    /// in a Route are invalid, the entire Route would be considered invalid. If only
+    /// a portion of Route Rules are invalid, implementations MUST set the
+    /// "PartiallyInvalid" condition for the Route.
     ///
     /// Conformance-levels at this level are defined based on the type of filter:
     ///
-    /// - ALL core filters MUST be supported by all implementations that support
-    ///   GRPCRoute.
+    /// - ALL core filters MUST be supported by all implementations.
     /// - Implementers are encouraged to support extended filters.
     /// - Implementation-specific custom filters have no API guarantees across
     ///   implementations.
@@ -336,7 +332,9 @@ pub struct GRPCRouteRules {
     /// Specifying the same filter multiple times is not supported unless explicitly
     /// indicated in the filter.
     ///
-    /// If an implementation can not support a combination of filters, it must clearly
+    /// All filters are expected to be compatible with each other except for the
+    /// URLRewrite and RequestRedirect filters, which may not be combined. If an
+    /// implementation can not support other combinations of filters, they must clearly
     /// document that limitation. In cases where incompatible or unsupported
     /// filters are specified and cause the `Accepted` condition to be set to status
     /// `False`, implementations may use the `IncompatibleFilters` reason to specify
@@ -344,45 +342,49 @@ pub struct GRPCRouteRules {
     ///
     /// Support: Core
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub filters: Option<Vec<GRPCRouteRulesFilters>>,
+    pub filters: Option<Vec<HTTPRouteRulesFilters>>,
     /// Matches define conditions used for matching the rule against incoming
-    /// gRPC requests. Each match is independent, i.e. this rule will be matched
+    /// HTTP requests. Each match is independent, i.e. this rule will be matched
     /// if **any** one of the matches is satisfied.
     ///
     /// For example, take the following matches configuration:
     ///
-    /// ```
+    /// ```text
     /// matches:
-    /// - method:
-    ///     service: foo.bar
+    /// - path:
+    ///     value: "/foo"
     ///   headers:
-    ///     values:
-    ///       version: 2
-    /// - method:
-    ///     service: foo.bar.v2
+    ///   - name: "version"
+    ///     value: "v2"
+    /// - path:
+    ///     value: "/v2/foo"
     /// ```
     ///
-    /// For a request to match against this rule, it MUST satisfy
+    /// For a request to match against this rule, a request must satisfy
     /// EITHER of the two conditions:
     ///
-    /// - service of foo.bar AND contains the header `version: 2`
-    /// - service of foo.bar.v2
+    /// - path prefixed with `/foo` AND contains the header `version: v2`
+    /// - path prefix of `/v2/foo`
     ///
-    /// See the documentation for GRPCRouteMatch on how to specify multiple
-    /// match conditions to be ANDed together.
+    /// See the documentation for HTTPRouteMatch on how to specify multiple
+    /// match conditions that should be ANDed together.
     ///
-    /// If no matches are specified, the implementation MUST match every gRPC request.
+    /// If no matches are specified, the default is a prefix
+    /// path match on "/", which has the effect of matching every
+    /// HTTP request.
     ///
-    /// Proxy or Load Balancer routing configuration generated from GRPCRoutes
-    /// MUST prioritize rules based on the following criteria, continuing on
-    /// ties. Merging MUST not be done between GRPCRoutes and HTTPRoutes.
-    /// Precedence MUST be given to the rule with the largest number of:
+    /// Proxy or Load Balancer routing configuration generated from HTTPRoutes
+    /// MUST prioritize matches based on the following criteria, continuing on
+    /// ties. Across all rules specified on applicable Routes, precedence must be
+    /// given to the match having:
     ///
-    /// * Characters in a matching non-wildcard hostname.
-    /// * Characters in a matching hostname.
-    /// * Characters in a matching service.
-    /// * Characters in a matching method.
-    /// * Header matches.
+    /// * "Exact" path match.
+    /// * "Prefix" path match with largest number of characters.
+    /// * Method match.
+    /// * Largest number of header matches.
+    /// * Largest number of query param matches.
+    ///
+    /// Note: The precedence of RegularExpression path matches are implementation-specific.
     ///
     /// If ties still exist across multiple Routes, matching precedence MUST be
     /// determined in order of the following criteria, continuing on ties:
@@ -391,32 +393,22 @@ pub struct GRPCRouteRules {
     /// * The Route appearing first in alphabetical order by
     ///   "{namespace}/{name}".
     ///
-    /// If ties still exist within the Route that has been given precedence,
-    /// matching precedence MUST be granted to the first matching rule meeting
-    /// the above criteria.
+    /// If ties still exist within an HTTPRoute, matching precedence MUST be granted
+    /// to the FIRST matching rule (in list order) with a match meeting the above
+    /// criteria.
+    ///
+    /// When no rules matching a request have been successfully attached to the
+    /// parent a request is coming from, a HTTP 404 status code MUST be returned.
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub matches: Option<Vec<GRPCRouteRulesMatches>>,
-    /// Name is the name of the route rule. This name MUST be unique within a Route if it is set.
+    pub matches: Option<Vec<HTTPRouteRulesMatches>>,
+    /// Timeouts defines the timeouts that can be configured for an HTTP request.
     ///
     /// Support: Extended
-    ///
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    /// SessionPersistence defines and configures session persistence
-    /// for the route rule.
-    ///
-    /// Support: Extended
-    ///
-    ///
-    #[serde(
-        default,
-        skip_serializing_if = "Option::is_none",
-        rename = "sessionPersistence"
-    )]
-    pub session_persistence: Option<GRPCRouteRulesSessionPersistence>,
+    pub timeouts: Option<HTTPRouteRulesTimeouts>,
 }
 
-/// GRPCBackendRef defines how a GRPCRoute forwards a gRPC request.
+/// HTTPBackendRef defines how a HTTPRoute forwards a HTTP request.
 ///
 /// Note that when a namespace different than the local namespace is specified, a
 /// ReferenceGrant object is required in the referent namespace to allow that
@@ -441,14 +433,14 @@ pub struct GRPCRouteRules {
 ///
 /// </gateway:experimental:description>
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesBackendRefs {
-    /// Filters defined at this level MUST be executed if and only if the
+pub struct HTTPRouteRulesBackendRefs {
+    /// Filters defined at this level should be executed if and only if the
     /// request is being forwarded to the backend defined here.
     ///
     /// Support: Implementation-specific (For broader support of filters, use the
-    /// Filters field in GRPCRouteRule.)
+    /// Filters field in HTTPRouteRule.)
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub filters: Option<Vec<GRPCRouteRulesBackendRefsFilters>>,
+    pub filters: Option<Vec<HTTPRouteRulesBackendRefsFilters>>,
     /// Group is the group of the referent. For example, "gateway.networking.k8s.io".
     /// When unspecified or empty string, core API group is inferred.
     #[serde(default, skip_serializing_if = "Option::is_none")]
@@ -506,28 +498,28 @@ pub struct GRPCRouteRulesBackendRefs {
     pub weight: Option<i32>,
 }
 
-/// GRPCRouteFilter defines processing steps that must be completed during the
-/// request or response lifecycle. GRPCRouteFilters are meant as an extension
+/// HTTPRouteFilter defines processing steps that must be completed during the
+/// request or response lifecycle. HTTPRouteFilters are meant as an extension
 /// point to express processing that may be done in Gateway implementations. Some
 /// examples include request or response modification, implementing
 /// authentication strategies, rate-limiting, and traffic shaping. API
 /// guarantee/conformance is defined based on the type of the filter.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesBackendRefsFilters {
+pub struct HTTPRouteRulesBackendRefsFilters {
     /// ExtensionRef is an optional, implementation-specific extension to the
     /// "filter" behavior.  For example, resource "myroutefilter" in group
     /// "networking.example.net"). ExtensionRef MUST NOT be used for core and
     /// extended filters.
     ///
-    /// Support: Implementation-specific
-    ///
     /// This filter can be used multiple times within the same rule.
+    ///
+    /// Support: Implementation-specific
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "extensionRef"
     )]
-    pub extension_ref: Option<GRPCRouteRulesBackendRefsFiltersExtensionRef>,
+    pub extension_ref: Option<HTTPRouteRulesBackendRefsFiltersExtensionRef>,
     /// RequestHeaderModifier defines a schema for a filter that modifies request
     /// headers.
     ///
@@ -537,7 +529,7 @@ pub struct GRPCRouteRulesBackendRefsFilters {
         skip_serializing_if = "Option::is_none",
         rename = "requestHeaderModifier"
     )]
-    pub request_header_modifier: Option<GRPCRouteRulesBackendRefsFiltersRequestHeaderModifier>,
+    pub request_header_modifier: Option<HTTPRouteRulesBackendRefsFiltersRequestHeaderModifier>,
     /// RequestMirror defines a schema for a filter that mirrors requests.
     /// Requests are sent to the specified destination, but responses from
     /// that destination are ignored.
@@ -554,7 +546,17 @@ pub struct GRPCRouteRulesBackendRefsFilters {
         skip_serializing_if = "Option::is_none",
         rename = "requestMirror"
     )]
-    pub request_mirror: Option<GRPCRouteRulesBackendRefsFiltersRequestMirror>,
+    pub request_mirror: Option<HTTPRouteRulesBackendRefsFiltersRequestMirror>,
+    /// RequestRedirect defines a schema for a filter that responds to the
+    /// request with an HTTP redirection.
+    ///
+    /// Support: Core
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "requestRedirect"
+    )]
+    pub request_redirect: Option<HTTPRouteRulesBackendRefsFiltersRequestRedirect>,
     /// ResponseHeaderModifier defines a schema for a filter that modifies response
     /// headers.
     ///
@@ -564,23 +566,24 @@ pub struct GRPCRouteRulesBackendRefsFilters {
         skip_serializing_if = "Option::is_none",
         rename = "responseHeaderModifier"
     )]
-    pub response_header_modifier: Option<GRPCRouteRulesBackendRefsFiltersResponseHeaderModifier>,
+    pub response_header_modifier: Option<HTTPRouteRulesBackendRefsFiltersResponseHeaderModifier>,
     /// Type identifies the type of filter to apply. As with other API fields,
     /// types are classified into three conformance levels:
     ///
     /// - Core: Filter types and their corresponding configuration defined by
     ///   "Support: Core" in this package, e.g. "RequestHeaderModifier". All
-    ///   implementations supporting GRPCRoute MUST support core filters.
+    ///   implementations must support core filters.
     ///
     /// - Extended: Filter types and their corresponding configuration defined by
     ///   "Support: Extended" in this package, e.g. "RequestMirror". Implementers
     ///   are encouraged to support extended filters.
     ///
-    /// - Implementation-specific: Filters that are defined and supported by specific vendors.
+    /// - Implementation-specific: Filters that are defined and supported by
+    ///   specific vendors.
     ///   In the future, filters showing convergence in behavior across multiple
     ///   implementations will be considered for inclusion in extended or core
     ///   conformance levels. Filter-specific configuration for such filters
-    ///   is specified using the ExtensionRef field. `Type` MUST be set to
+    ///   is specified using the ExtensionRef field. `Type` should be set to
     ///   "ExtensionRef" for custom filters.
     ///
     /// Implementers are encouraged to define custom implementation types to
@@ -590,9 +593,23 @@ pub struct GRPCRouteRulesBackendRefsFilters {
     /// MUST NOT be skipped. Instead, requests that would have been processed by
     /// that filter MUST receive a HTTP error response.
     ///
+    /// Note that values may be added to this enum, implementations
+    /// must ensure that unknown values will not cause a crash.
     ///
+    /// Unknown values here must result in the implementation setting the
+    /// Accepted Condition for the Route to `status: False`, with a
+    /// Reason of `UnsupportedValue`.
     #[serde(rename = "type")]
-    pub r#type: GRPCRouteRulesBackendRefsFiltersType,
+    pub r#type: HTTPRouteRulesBackendRefsFiltersType,
+    /// URLRewrite defines a schema for a filter that modifies a request during forwarding.
+    ///
+    /// Support: Extended
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "urlRewrite"
+    )]
+    pub url_rewrite: Option<HTTPRouteRulesBackendRefsFiltersUrlRewrite>,
 }
 
 /// ExtensionRef is an optional, implementation-specific extension to the
@@ -600,11 +617,11 @@ pub struct GRPCRouteRulesBackendRefsFilters {
 /// "networking.example.net"). ExtensionRef MUST NOT be used for core and
 /// extended filters.
 ///
-/// Support: Implementation-specific
-///
 /// This filter can be used multiple times within the same rule.
+///
+/// Support: Implementation-specific
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesBackendRefsFiltersExtensionRef {
+pub struct HTTPRouteRulesBackendRefsFiltersExtensionRef {
     /// Group is the group of the referent. For example, "gateway.networking.k8s.io".
     /// When unspecified or empty string, core API group is inferred.
     pub group: String,
@@ -619,7 +636,7 @@ pub struct GRPCRouteRulesBackendRefsFiltersExtensionRef {
 ///
 /// Support: Core
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesBackendRefsFiltersRequestHeaderModifier {
+pub struct HTTPRouteRulesBackendRefsFiltersRequestHeaderModifier {
     /// Add adds the given header(s) (name, value) to the request
     /// before the action. It appends to any existing values associated
     /// with the header name.
@@ -637,7 +654,7 @@ pub struct GRPCRouteRulesBackendRefsFiltersRequestHeaderModifier {
     ///   GET /foo HTTP/1.1
     ///   my-header: foo,bar,baz
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub add: Option<Vec<GRPCRouteRulesBackendRefsFiltersRequestHeaderModifierAdd>>,
+    pub add: Option<Vec<HTTPRouteRulesBackendRefsFiltersRequestHeaderModifierAdd>>,
     /// Remove the given header(s) from the HTTP request before the action. The
     /// value of Remove is a list of HTTP header names. Note that the header
     /// names are case-insensitive (see
@@ -673,12 +690,12 @@ pub struct GRPCRouteRulesBackendRefsFiltersRequestHeaderModifier {
     ///   GET /foo HTTP/1.1
     ///   my-header: bar
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub set: Option<Vec<GRPCRouteRulesBackendRefsFiltersRequestHeaderModifierSet>>,
+    pub set: Option<Vec<HTTPRouteRulesBackendRefsFiltersRequestHeaderModifierSet>>,
 }
 
 /// HTTPHeader represents an HTTP Header name and value as defined by RFC 7230.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesBackendRefsFiltersRequestHeaderModifierAdd {
+pub struct HTTPRouteRulesBackendRefsFiltersRequestHeaderModifierAdd {
     /// Name is the name of the HTTP Header to be matched. Name matching MUST be
     /// case insensitive. (See https://tools.ietf.org/html/rfc7230#section-3.2).
     ///
@@ -694,7 +711,7 @@ pub struct GRPCRouteRulesBackendRefsFiltersRequestHeaderModifierAdd {
 
 /// HTTPHeader represents an HTTP Header name and value as defined by RFC 7230.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesBackendRefsFiltersRequestHeaderModifierSet {
+pub struct HTTPRouteRulesBackendRefsFiltersRequestHeaderModifierSet {
     /// Name is the name of the HTTP Header to be matched. Name matching MUST be
     /// case insensitive. (See https://tools.ietf.org/html/rfc7230#section-3.2).
     ///
@@ -720,7 +737,7 @@ pub struct GRPCRouteRulesBackendRefsFiltersRequestHeaderModifierSet {
 ///
 ///
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesBackendRefsFiltersRequestMirror {
+pub struct HTTPRouteRulesBackendRefsFiltersRequestMirror {
     /// BackendRef references a resource where mirrored requests are sent.
     ///
     /// Mirrored requests must be sent only to a single destination endpoint
@@ -745,26 +762,7 @@ pub struct GRPCRouteRulesBackendRefsFiltersRequestMirror {
     ///
     /// Support: Implementation-specific for any other resource
     #[serde(rename = "backendRef")]
-    pub backend_ref: GRPCRouteRulesBackendRefsFiltersRequestMirrorBackendRef,
-    /// Fraction represents the fraction of requests that should be
-    /// mirrored to BackendRef.
-    ///
-    /// Only one of Fraction or Percent may be specified. If neither field
-    /// is specified, 100% of requests will be mirrored.
-    ///
-    ///
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub fraction: Option<GRPCRouteRulesBackendRefsFiltersRequestMirrorFraction>,
-    /// Percent represents the percentage of requests that should be
-    /// mirrored to BackendRef. Its minimum value is 0 (indicating 0% of
-    /// requests) and its maximum value is 100 (indicating 100% of requests).
-    ///
-    /// Only one of Fraction or Percent may be specified. If neither field
-    /// is specified, 100% of requests will be mirrored.
-    ///
-    ///
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub percent: Option<i32>,
+    pub backend_ref: HTTPRouteRulesBackendRefsFiltersRequestMirrorBackendRef,
 }
 
 /// BackendRef references a resource where mirrored requests are sent.
@@ -791,7 +789,7 @@ pub struct GRPCRouteRulesBackendRefsFiltersRequestMirror {
 ///
 /// Support: Implementation-specific for any other resource
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesBackendRefsFiltersRequestMirrorBackendRef {
+pub struct HTTPRouteRulesBackendRefsFiltersRequestMirrorBackendRef {
     /// Group is the group of the referent. For example, "gateway.networking.k8s.io".
     /// When unspecified or empty string, core API group is inferred.
     #[serde(default, skip_serializing_if = "Option::is_none")]
@@ -834,18 +832,167 @@ pub struct GRPCRouteRulesBackendRefsFiltersRequestMirrorBackendRef {
     pub port: Option<i32>,
 }
 
-/// Fraction represents the fraction of requests that should be
-/// mirrored to BackendRef.
+/// RequestRedirect defines a schema for a filter that responds to the
+/// request with an HTTP redirection.
 ///
-/// Only one of Fraction or Percent may be specified. If neither field
-/// is specified, 100% of requests will be mirrored.
-///
-///
+/// Support: Core
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesBackendRefsFiltersRequestMirrorFraction {
+pub struct HTTPRouteRulesBackendRefsFiltersRequestRedirect {
+    /// Hostname is the hostname to be used in the value of the `Location`
+    /// header in the response.
+    /// When empty, the hostname in the `Host` header of the request is used.
+    ///
+    /// Support: Core
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub denominator: Option<i32>,
-    pub numerator: i32,
+    pub hostname: Option<String>,
+    /// Path defines parameters used to modify the path of the incoming request.
+    /// The modified path is then used to construct the `Location` header. When
+    /// empty, the request path is used as-is.
+    ///
+    /// Support: Extended
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub path: Option<HTTPRouteRulesBackendRefsFiltersRequestRedirectPath>,
+    /// Port is the port to be used in the value of the `Location`
+    /// header in the response.
+    ///
+    /// If no port is specified, the redirect port MUST be derived using the
+    /// following rules:
+    ///
+    /// * If redirect scheme is not-empty, the redirect port MUST be the well-known
+    ///   port associated with the redirect scheme. Specifically "http" to port 80
+    ///   and "https" to port 443. If the redirect scheme does not have a
+    ///   well-known port, the listener port of the Gateway SHOULD be used.
+    /// * If redirect scheme is empty, the redirect port MUST be the Gateway
+    ///   Listener port.
+    ///
+    /// Implementations SHOULD NOT add the port number in the 'Location'
+    /// header in the following cases:
+    ///
+    /// * A Location header that will use HTTP (whether that is determined via
+    ///   the Listener protocol or the Scheme field) _and_ use port 80.
+    /// * A Location header that will use HTTPS (whether that is determined via
+    ///   the Listener protocol or the Scheme field) _and_ use port 443.
+    ///
+    /// Support: Extended
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub port: Option<i32>,
+    /// Scheme is the scheme to be used in the value of the `Location` header in
+    /// the response. When empty, the scheme of the request is used.
+    ///
+    /// Scheme redirects can affect the port of the redirect, for more information,
+    /// refer to the documentation for the port field of this filter.
+    ///
+    /// Note that values may be added to this enum, implementations
+    /// must ensure that unknown values will not cause a crash.
+    ///
+    /// Unknown values here must result in the implementation setting the
+    /// Accepted Condition for the Route to `status: False`, with a
+    /// Reason of `UnsupportedValue`.
+    ///
+    /// Support: Extended
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub scheme: Option<HTTPRouteRulesBackendRefsFiltersRequestRedirectScheme>,
+    /// StatusCode is the HTTP status code to be used in response.
+    ///
+    /// Note that values may be added to this enum, implementations
+    /// must ensure that unknown values will not cause a crash.
+    ///
+    /// Unknown values here must result in the implementation setting the
+    /// Accepted Condition for the Route to `status: False`, with a
+    /// Reason of `UnsupportedValue`.
+    ///
+    /// Support: Core
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "statusCode"
+    )]
+    pub status_code: Option<i64>,
+}
+
+/// Path defines parameters used to modify the path of the incoming request.
+/// The modified path is then used to construct the `Location` header. When
+/// empty, the request path is used as-is.
+///
+/// Support: Extended
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
+pub struct HTTPRouteRulesBackendRefsFiltersRequestRedirectPath {
+    /// ReplaceFullPath specifies the value with which to replace the full path
+    /// of a request during a rewrite or redirect.
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "replaceFullPath"
+    )]
+    pub replace_full_path: Option<String>,
+    /// ReplacePrefixMatch specifies the value with which to replace the prefix
+    /// match of a request during a rewrite or redirect. For example, a request
+    /// to "/foo/bar" with a prefix match of "/foo" and a ReplacePrefixMatch
+    /// of "/xyz" would be modified to "/xyz/bar".
+    ///
+    /// Note that this matches the behavior of the PathPrefix match type. This
+    /// matches full path elements. A path element refers to the list of labels
+    /// in the path split by the `/` separator. When specified, a trailing `/` is
+    /// ignored. For example, the paths `/abc`, `/abc/`, and `/abc/def` would all
+    /// match the prefix `/abc`, but the path `/abcd` would not.
+    ///
+    /// ReplacePrefixMatch is only compatible with a `PathPrefix` HTTPRouteMatch.
+    /// Using any other HTTPRouteMatch type on the same HTTPRouteRule will result in
+    /// the implementation setting the Accepted Condition for the Route to `status: False`.
+    ///
+    /// Request Path | Prefix Match | Replace Prefix | Modified Path
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "replacePrefixMatch"
+    )]
+    pub replace_prefix_match: Option<String>,
+    /// Type defines the type of path modifier. Additional types may be
+    /// added in a future release of the API.
+    ///
+    /// Note that values may be added to this enum, implementations
+    /// must ensure that unknown values will not cause a crash.
+    ///
+    /// Unknown values here must result in the implementation setting the
+    /// Accepted Condition for the Route to `status: False`, with a
+    /// Reason of `UnsupportedValue`.
+    #[serde(rename = "type")]
+    pub r#type: HTTPRouteRulesBackendRefsFiltersRequestRedirectPathType,
+}
+
+/// Path defines parameters used to modify the path of the incoming request.
+/// The modified path is then used to construct the `Location` header. When
+/// empty, the request path is used as-is.
+///
+/// Support: Extended
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, PartialEq)]
+pub enum HTTPRouteRulesBackendRefsFiltersRequestRedirectPathType {
+    ReplaceFullPath,
+    ReplacePrefixMatch,
+}
+
+/// RequestRedirect defines a schema for a filter that responds to the
+/// request with an HTTP redirection.
+///
+/// Support: Core
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, PartialEq)]
+pub enum HTTPRouteRulesBackendRefsFiltersRequestRedirectScheme {
+    #[serde(rename = "http")]
+    Http,
+    #[serde(rename = "https")]
+    Https,
+}
+
+/// RequestRedirect defines a schema for a filter that responds to the
+/// request with an HTTP redirection.
+///
+/// Support: Core
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, PartialEq)]
+pub enum HTTPRouteRulesBackendRefsFiltersRequestRedirectStatusCode {
+    #[serde(rename = "301")]
+    r#_301,
+    #[serde(rename = "302")]
+    r#_302,
 }
 
 /// ResponseHeaderModifier defines a schema for a filter that modifies response
@@ -853,7 +1000,7 @@ pub struct GRPCRouteRulesBackendRefsFiltersRequestMirrorFraction {
 ///
 /// Support: Extended
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesBackendRefsFiltersResponseHeaderModifier {
+pub struct HTTPRouteRulesBackendRefsFiltersResponseHeaderModifier {
     /// Add adds the given header(s) (name, value) to the request
     /// before the action. It appends to any existing values associated
     /// with the header name.
@@ -871,7 +1018,7 @@ pub struct GRPCRouteRulesBackendRefsFiltersResponseHeaderModifier {
     ///   GET /foo HTTP/1.1
     ///   my-header: foo,bar,baz
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub add: Option<Vec<GRPCRouteRulesBackendRefsFiltersResponseHeaderModifierAdd>>,
+    pub add: Option<Vec<HTTPRouteRulesBackendRefsFiltersResponseHeaderModifierAdd>>,
     /// Remove the given header(s) from the HTTP request before the action. The
     /// value of Remove is a list of HTTP header names. Note that the header
     /// names are case-insensitive (see
@@ -907,12 +1054,12 @@ pub struct GRPCRouteRulesBackendRefsFiltersResponseHeaderModifier {
     ///   GET /foo HTTP/1.1
     ///   my-header: bar
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub set: Option<Vec<GRPCRouteRulesBackendRefsFiltersResponseHeaderModifierSet>>,
+    pub set: Option<Vec<HTTPRouteRulesBackendRefsFiltersResponseHeaderModifierSet>>,
 }
 
 /// HTTPHeader represents an HTTP Header name and value as defined by RFC 7230.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesBackendRefsFiltersResponseHeaderModifierAdd {
+pub struct HTTPRouteRulesBackendRefsFiltersResponseHeaderModifierAdd {
     /// Name is the name of the HTTP Header to be matched. Name matching MUST be
     /// case insensitive. (See https://tools.ietf.org/html/rfc7230#section-3.2).
     ///
@@ -928,7 +1075,7 @@ pub struct GRPCRouteRulesBackendRefsFiltersResponseHeaderModifierAdd {
 
 /// HTTPHeader represents an HTTP Header name and value as defined by RFC 7230.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesBackendRefsFiltersResponseHeaderModifierSet {
+pub struct HTTPRouteRulesBackendRefsFiltersResponseHeaderModifierSet {
     /// Name is the name of the HTTP Header to be matched. Name matching MUST be
     /// case insensitive. (See https://tools.ietf.org/html/rfc7230#section-3.2).
     ///
@@ -942,42 +1089,120 @@ pub struct GRPCRouteRulesBackendRefsFiltersResponseHeaderModifierSet {
     pub value: String,
 }
 
-/// GRPCRouteFilter defines processing steps that must be completed during the
-/// request or response lifecycle. GRPCRouteFilters are meant as an extension
+/// HTTPRouteFilter defines processing steps that must be completed during the
+/// request or response lifecycle. HTTPRouteFilters are meant as an extension
 /// point to express processing that may be done in Gateway implementations. Some
 /// examples include request or response modification, implementing
 /// authentication strategies, rate-limiting, and traffic shaping. API
 /// guarantee/conformance is defined based on the type of the filter.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, PartialEq)]
-pub enum GRPCRouteRulesBackendRefsFiltersType {
-    ResponseHeaderModifier,
+pub enum HTTPRouteRulesBackendRefsFiltersType {
     RequestHeaderModifier,
+    ResponseHeaderModifier,
     RequestMirror,
+    RequestRedirect,
+    #[serde(rename = "URLRewrite")]
+    UrlRewrite,
     ExtensionRef,
 }
 
-/// GRPCRouteFilter defines processing steps that must be completed during the
-/// request or response lifecycle. GRPCRouteFilters are meant as an extension
+/// URLRewrite defines a schema for a filter that modifies a request during forwarding.
+///
+/// Support: Extended
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
+pub struct HTTPRouteRulesBackendRefsFiltersUrlRewrite {
+    /// Hostname is the value to be used to replace the Host header value during
+    /// forwarding.
+    ///
+    /// Support: Extended
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub hostname: Option<String>,
+    /// Path defines a path rewrite.
+    ///
+    /// Support: Extended
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub path: Option<HTTPRouteRulesBackendRefsFiltersUrlRewritePath>,
+}
+
+/// Path defines a path rewrite.
+///
+/// Support: Extended
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
+pub struct HTTPRouteRulesBackendRefsFiltersUrlRewritePath {
+    /// ReplaceFullPath specifies the value with which to replace the full path
+    /// of a request during a rewrite or redirect.
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "replaceFullPath"
+    )]
+    pub replace_full_path: Option<String>,
+    /// ReplacePrefixMatch specifies the value with which to replace the prefix
+    /// match of a request during a rewrite or redirect. For example, a request
+    /// to "/foo/bar" with a prefix match of "/foo" and a ReplacePrefixMatch
+    /// of "/xyz" would be modified to "/xyz/bar".
+    ///
+    /// Note that this matches the behavior of the PathPrefix match type. This
+    /// matches full path elements. A path element refers to the list of labels
+    /// in the path split by the `/` separator. When specified, a trailing `/` is
+    /// ignored. For example, the paths `/abc`, `/abc/`, and `/abc/def` would all
+    /// match the prefix `/abc`, but the path `/abcd` would not.
+    ///
+    /// ReplacePrefixMatch is only compatible with a `PathPrefix` HTTPRouteMatch.
+    /// Using any other HTTPRouteMatch type on the same HTTPRouteRule will result in
+    /// the implementation setting the Accepted Condition for the Route to `status: False`.
+    ///
+    /// Request Path | Prefix Match | Replace Prefix | Modified Path
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "replacePrefixMatch"
+    )]
+    pub replace_prefix_match: Option<String>,
+    /// Type defines the type of path modifier. Additional types may be
+    /// added in a future release of the API.
+    ///
+    /// Note that values may be added to this enum, implementations
+    /// must ensure that unknown values will not cause a crash.
+    ///
+    /// Unknown values here must result in the implementation setting the
+    /// Accepted Condition for the Route to `status: False`, with a
+    /// Reason of `UnsupportedValue`.
+    #[serde(rename = "type")]
+    pub r#type: HTTPRouteRulesBackendRefsFiltersUrlRewritePathType,
+}
+
+/// Path defines a path rewrite.
+///
+/// Support: Extended
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, PartialEq)]
+pub enum HTTPRouteRulesBackendRefsFiltersUrlRewritePathType {
+    ReplaceFullPath,
+    ReplacePrefixMatch,
+}
+
+/// HTTPRouteFilter defines processing steps that must be completed during the
+/// request or response lifecycle. HTTPRouteFilters are meant as an extension
 /// point to express processing that may be done in Gateway implementations. Some
 /// examples include request or response modification, implementing
 /// authentication strategies, rate-limiting, and traffic shaping. API
 /// guarantee/conformance is defined based on the type of the filter.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesFilters {
+pub struct HTTPRouteRulesFilters {
     /// ExtensionRef is an optional, implementation-specific extension to the
     /// "filter" behavior.  For example, resource "myroutefilter" in group
     /// "networking.example.net"). ExtensionRef MUST NOT be used for core and
     /// extended filters.
     ///
-    /// Support: Implementation-specific
-    ///
     /// This filter can be used multiple times within the same rule.
+    ///
+    /// Support: Implementation-specific
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "extensionRef"
     )]
-    pub extension_ref: Option<GRPCRouteRulesFiltersExtensionRef>,
+    pub extension_ref: Option<HTTPRouteRulesFiltersExtensionRef>,
     /// RequestHeaderModifier defines a schema for a filter that modifies request
     /// headers.
     ///
@@ -987,7 +1212,7 @@ pub struct GRPCRouteRulesFilters {
         skip_serializing_if = "Option::is_none",
         rename = "requestHeaderModifier"
     )]
-    pub request_header_modifier: Option<GRPCRouteRulesFiltersRequestHeaderModifier>,
+    pub request_header_modifier: Option<HTTPRouteRulesFiltersRequestHeaderModifier>,
     /// RequestMirror defines a schema for a filter that mirrors requests.
     /// Requests are sent to the specified destination, but responses from
     /// that destination are ignored.
@@ -1004,7 +1229,17 @@ pub struct GRPCRouteRulesFilters {
         skip_serializing_if = "Option::is_none",
         rename = "requestMirror"
     )]
-    pub request_mirror: Option<GRPCRouteRulesFiltersRequestMirror>,
+    pub request_mirror: Option<HTTPRouteRulesFiltersRequestMirror>,
+    /// RequestRedirect defines a schema for a filter that responds to the
+    /// request with an HTTP redirection.
+    ///
+    /// Support: Core
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "requestRedirect"
+    )]
+    pub request_redirect: Option<HTTPRouteRulesFiltersRequestRedirect>,
     /// ResponseHeaderModifier defines a schema for a filter that modifies response
     /// headers.
     ///
@@ -1014,23 +1249,24 @@ pub struct GRPCRouteRulesFilters {
         skip_serializing_if = "Option::is_none",
         rename = "responseHeaderModifier"
     )]
-    pub response_header_modifier: Option<GRPCRouteRulesFiltersResponseHeaderModifier>,
+    pub response_header_modifier: Option<HTTPRouteRulesFiltersResponseHeaderModifier>,
     /// Type identifies the type of filter to apply. As with other API fields,
     /// types are classified into three conformance levels:
     ///
     /// - Core: Filter types and their corresponding configuration defined by
     ///   "Support: Core" in this package, e.g. "RequestHeaderModifier". All
-    ///   implementations supporting GRPCRoute MUST support core filters.
+    ///   implementations must support core filters.
     ///
     /// - Extended: Filter types and their corresponding configuration defined by
     ///   "Support: Extended" in this package, e.g. "RequestMirror". Implementers
     ///   are encouraged to support extended filters.
     ///
-    /// - Implementation-specific: Filters that are defined and supported by specific vendors.
+    /// - Implementation-specific: Filters that are defined and supported by
+    ///   specific vendors.
     ///   In the future, filters showing convergence in behavior across multiple
     ///   implementations will be considered for inclusion in extended or core
     ///   conformance levels. Filter-specific configuration for such filters
-    ///   is specified using the ExtensionRef field. `Type` MUST be set to
+    ///   is specified using the ExtensionRef field. `Type` should be set to
     ///   "ExtensionRef" for custom filters.
     ///
     /// Implementers are encouraged to define custom implementation types to
@@ -1040,9 +1276,23 @@ pub struct GRPCRouteRulesFilters {
     /// MUST NOT be skipped. Instead, requests that would have been processed by
     /// that filter MUST receive a HTTP error response.
     ///
+    /// Note that values may be added to this enum, implementations
+    /// must ensure that unknown values will not cause a crash.
     ///
+    /// Unknown values here must result in the implementation setting the
+    /// Accepted Condition for the Route to `status: False`, with a
+    /// Reason of `UnsupportedValue`.
     #[serde(rename = "type")]
-    pub r#type: GRPCRouteRulesFiltersType,
+    pub r#type: HTTPRouteRulesFiltersType,
+    /// URLRewrite defines a schema for a filter that modifies a request during forwarding.
+    ///
+    /// Support: Extended
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "urlRewrite"
+    )]
+    pub url_rewrite: Option<HTTPRouteRulesFiltersUrlRewrite>,
 }
 
 /// ExtensionRef is an optional, implementation-specific extension to the
@@ -1050,11 +1300,11 @@ pub struct GRPCRouteRulesFilters {
 /// "networking.example.net"). ExtensionRef MUST NOT be used for core and
 /// extended filters.
 ///
-/// Support: Implementation-specific
-///
 /// This filter can be used multiple times within the same rule.
+///
+/// Support: Implementation-specific
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesFiltersExtensionRef {
+pub struct HTTPRouteRulesFiltersExtensionRef {
     /// Group is the group of the referent. For example, "gateway.networking.k8s.io".
     /// When unspecified or empty string, core API group is inferred.
     pub group: String,
@@ -1069,7 +1319,7 @@ pub struct GRPCRouteRulesFiltersExtensionRef {
 ///
 /// Support: Core
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesFiltersRequestHeaderModifier {
+pub struct HTTPRouteRulesFiltersRequestHeaderModifier {
     /// Add adds the given header(s) (name, value) to the request
     /// before the action. It appends to any existing values associated
     /// with the header name.
@@ -1087,7 +1337,7 @@ pub struct GRPCRouteRulesFiltersRequestHeaderModifier {
     ///   GET /foo HTTP/1.1
     ///   my-header: foo,bar,baz
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub add: Option<Vec<GRPCRouteRulesFiltersRequestHeaderModifierAdd>>,
+    pub add: Option<Vec<HTTPRouteRulesFiltersRequestHeaderModifierAdd>>,
     /// Remove the given header(s) from the HTTP request before the action. The
     /// value of Remove is a list of HTTP header names. Note that the header
     /// names are case-insensitive (see
@@ -1123,12 +1373,12 @@ pub struct GRPCRouteRulesFiltersRequestHeaderModifier {
     ///   GET /foo HTTP/1.1
     ///   my-header: bar
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub set: Option<Vec<GRPCRouteRulesFiltersRequestHeaderModifierSet>>,
+    pub set: Option<Vec<HTTPRouteRulesFiltersRequestHeaderModifierSet>>,
 }
 
 /// HTTPHeader represents an HTTP Header name and value as defined by RFC 7230.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesFiltersRequestHeaderModifierAdd {
+pub struct HTTPRouteRulesFiltersRequestHeaderModifierAdd {
     /// Name is the name of the HTTP Header to be matched. Name matching MUST be
     /// case insensitive. (See https://tools.ietf.org/html/rfc7230#section-3.2).
     ///
@@ -1144,7 +1394,7 @@ pub struct GRPCRouteRulesFiltersRequestHeaderModifierAdd {
 
 /// HTTPHeader represents an HTTP Header name and value as defined by RFC 7230.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesFiltersRequestHeaderModifierSet {
+pub struct HTTPRouteRulesFiltersRequestHeaderModifierSet {
     /// Name is the name of the HTTP Header to be matched. Name matching MUST be
     /// case insensitive. (See https://tools.ietf.org/html/rfc7230#section-3.2).
     ///
@@ -1170,7 +1420,7 @@ pub struct GRPCRouteRulesFiltersRequestHeaderModifierSet {
 ///
 ///
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesFiltersRequestMirror {
+pub struct HTTPRouteRulesFiltersRequestMirror {
     /// BackendRef references a resource where mirrored requests are sent.
     ///
     /// Mirrored requests must be sent only to a single destination endpoint
@@ -1195,26 +1445,7 @@ pub struct GRPCRouteRulesFiltersRequestMirror {
     ///
     /// Support: Implementation-specific for any other resource
     #[serde(rename = "backendRef")]
-    pub backend_ref: GRPCRouteRulesFiltersRequestMirrorBackendRef,
-    /// Fraction represents the fraction of requests that should be
-    /// mirrored to BackendRef.
-    ///
-    /// Only one of Fraction or Percent may be specified. If neither field
-    /// is specified, 100% of requests will be mirrored.
-    ///
-    ///
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub fraction: Option<GRPCRouteRulesFiltersRequestMirrorFraction>,
-    /// Percent represents the percentage of requests that should be
-    /// mirrored to BackendRef. Its minimum value is 0 (indicating 0% of
-    /// requests) and its maximum value is 100 (indicating 100% of requests).
-    ///
-    /// Only one of Fraction or Percent may be specified. If neither field
-    /// is specified, 100% of requests will be mirrored.
-    ///
-    ///
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub percent: Option<i32>,
+    pub backend_ref: HTTPRouteRulesFiltersRequestMirrorBackendRef,
 }
 
 /// BackendRef references a resource where mirrored requests are sent.
@@ -1241,7 +1472,7 @@ pub struct GRPCRouteRulesFiltersRequestMirror {
 ///
 /// Support: Implementation-specific for any other resource
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesFiltersRequestMirrorBackendRef {
+pub struct HTTPRouteRulesFiltersRequestMirrorBackendRef {
     /// Group is the group of the referent. For example, "gateway.networking.k8s.io".
     /// When unspecified or empty string, core API group is inferred.
     #[serde(default, skip_serializing_if = "Option::is_none")]
@@ -1284,18 +1515,167 @@ pub struct GRPCRouteRulesFiltersRequestMirrorBackendRef {
     pub port: Option<i32>,
 }
 
-/// Fraction represents the fraction of requests that should be
-/// mirrored to BackendRef.
+/// RequestRedirect defines a schema for a filter that responds to the
+/// request with an HTTP redirection.
 ///
-/// Only one of Fraction or Percent may be specified. If neither field
-/// is specified, 100% of requests will be mirrored.
-///
-///
+/// Support: Core
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesFiltersRequestMirrorFraction {
+pub struct HTTPRouteRulesFiltersRequestRedirect {
+    /// Hostname is the hostname to be used in the value of the `Location`
+    /// header in the response.
+    /// When empty, the hostname in the `Host` header of the request is used.
+    ///
+    /// Support: Core
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub denominator: Option<i32>,
-    pub numerator: i32,
+    pub hostname: Option<String>,
+    /// Path defines parameters used to modify the path of the incoming request.
+    /// The modified path is then used to construct the `Location` header. When
+    /// empty, the request path is used as-is.
+    ///
+    /// Support: Extended
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub path: Option<HTTPRouteRulesFiltersRequestRedirectPath>,
+    /// Port is the port to be used in the value of the `Location`
+    /// header in the response.
+    ///
+    /// If no port is specified, the redirect port MUST be derived using the
+    /// following rules:
+    ///
+    /// * If redirect scheme is not-empty, the redirect port MUST be the well-known
+    ///   port associated with the redirect scheme. Specifically "http" to port 80
+    ///   and "https" to port 443. If the redirect scheme does not have a
+    ///   well-known port, the listener port of the Gateway SHOULD be used.
+    /// * If redirect scheme is empty, the redirect port MUST be the Gateway
+    ///   Listener port.
+    ///
+    /// Implementations SHOULD NOT add the port number in the 'Location'
+    /// header in the following cases:
+    ///
+    /// * A Location header that will use HTTP (whether that is determined via
+    ///   the Listener protocol or the Scheme field) _and_ use port 80.
+    /// * A Location header that will use HTTPS (whether that is determined via
+    ///   the Listener protocol or the Scheme field) _and_ use port 443.
+    ///
+    /// Support: Extended
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub port: Option<i32>,
+    /// Scheme is the scheme to be used in the value of the `Location` header in
+    /// the response. When empty, the scheme of the request is used.
+    ///
+    /// Scheme redirects can affect the port of the redirect, for more information,
+    /// refer to the documentation for the port field of this filter.
+    ///
+    /// Note that values may be added to this enum, implementations
+    /// must ensure that unknown values will not cause a crash.
+    ///
+    /// Unknown values here must result in the implementation setting the
+    /// Accepted Condition for the Route to `status: False`, with a
+    /// Reason of `UnsupportedValue`.
+    ///
+    /// Support: Extended
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub scheme: Option<HTTPRouteRulesFiltersRequestRedirectScheme>,
+    /// StatusCode is the HTTP status code to be used in response.
+    ///
+    /// Note that values may be added to this enum, implementations
+    /// must ensure that unknown values will not cause a crash.
+    ///
+    /// Unknown values here must result in the implementation setting the
+    /// Accepted Condition for the Route to `status: False`, with a
+    /// Reason of `UnsupportedValue`.
+    ///
+    /// Support: Core
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "statusCode"
+    )]
+    pub status_code: Option<i64>,
+}
+
+/// Path defines parameters used to modify the path of the incoming request.
+/// The modified path is then used to construct the `Location` header. When
+/// empty, the request path is used as-is.
+///
+/// Support: Extended
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
+pub struct HTTPRouteRulesFiltersRequestRedirectPath {
+    /// ReplaceFullPath specifies the value with which to replace the full path
+    /// of a request during a rewrite or redirect.
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "replaceFullPath"
+    )]
+    pub replace_full_path: Option<String>,
+    /// ReplacePrefixMatch specifies the value with which to replace the prefix
+    /// match of a request during a rewrite or redirect. For example, a request
+    /// to "/foo/bar" with a prefix match of "/foo" and a ReplacePrefixMatch
+    /// of "/xyz" would be modified to "/xyz/bar".
+    ///
+    /// Note that this matches the behavior of the PathPrefix match type. This
+    /// matches full path elements. A path element refers to the list of labels
+    /// in the path split by the `/` separator. When specified, a trailing `/` is
+    /// ignored. For example, the paths `/abc`, `/abc/`, and `/abc/def` would all
+    /// match the prefix `/abc`, but the path `/abcd` would not.
+    ///
+    /// ReplacePrefixMatch is only compatible with a `PathPrefix` HTTPRouteMatch.
+    /// Using any other HTTPRouteMatch type on the same HTTPRouteRule will result in
+    /// the implementation setting the Accepted Condition for the Route to `status: False`.
+    ///
+    /// Request Path | Prefix Match | Replace Prefix | Modified Path
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "replacePrefixMatch"
+    )]
+    pub replace_prefix_match: Option<String>,
+    /// Type defines the type of path modifier. Additional types may be
+    /// added in a future release of the API.
+    ///
+    /// Note that values may be added to this enum, implementations
+    /// must ensure that unknown values will not cause a crash.
+    ///
+    /// Unknown values here must result in the implementation setting the
+    /// Accepted Condition for the Route to `status: False`, with a
+    /// Reason of `UnsupportedValue`.
+    #[serde(rename = "type")]
+    pub r#type: HTTPRouteRulesFiltersRequestRedirectPathType,
+}
+
+/// Path defines parameters used to modify the path of the incoming request.
+/// The modified path is then used to construct the `Location` header. When
+/// empty, the request path is used as-is.
+///
+/// Support: Extended
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, PartialEq)]
+pub enum HTTPRouteRulesFiltersRequestRedirectPathType {
+    ReplaceFullPath,
+    ReplacePrefixMatch,
+}
+
+/// RequestRedirect defines a schema for a filter that responds to the
+/// request with an HTTP redirection.
+///
+/// Support: Core
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, PartialEq)]
+pub enum HTTPRouteRulesFiltersRequestRedirectScheme {
+    #[serde(rename = "http")]
+    Http,
+    #[serde(rename = "https")]
+    Https,
+}
+
+/// RequestRedirect defines a schema for a filter that responds to the
+/// request with an HTTP redirection.
+///
+/// Support: Core
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, PartialEq)]
+pub enum HTTPRouteRulesFiltersRequestRedirectStatusCode {
+    #[serde(rename = "301")]
+    r#_301,
+    #[serde(rename = "302")]
+    r#_302,
 }
 
 /// ResponseHeaderModifier defines a schema for a filter that modifies response
@@ -1303,7 +1683,7 @@ pub struct GRPCRouteRulesFiltersRequestMirrorFraction {
 ///
 /// Support: Extended
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesFiltersResponseHeaderModifier {
+pub struct HTTPRouteRulesFiltersResponseHeaderModifier {
     /// Add adds the given header(s) (name, value) to the request
     /// before the action. It appends to any existing values associated
     /// with the header name.
@@ -1321,7 +1701,7 @@ pub struct GRPCRouteRulesFiltersResponseHeaderModifier {
     ///   GET /foo HTTP/1.1
     ///   my-header: foo,bar,baz
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub add: Option<Vec<GRPCRouteRulesFiltersResponseHeaderModifierAdd>>,
+    pub add: Option<Vec<HTTPRouteRulesFiltersResponseHeaderModifierAdd>>,
     /// Remove the given header(s) from the HTTP request before the action. The
     /// value of Remove is a list of HTTP header names. Note that the header
     /// names are case-insensitive (see
@@ -1357,12 +1737,12 @@ pub struct GRPCRouteRulesFiltersResponseHeaderModifier {
     ///   GET /foo HTTP/1.1
     ///   my-header: bar
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub set: Option<Vec<GRPCRouteRulesFiltersResponseHeaderModifierSet>>,
+    pub set: Option<Vec<HTTPRouteRulesFiltersResponseHeaderModifierSet>>,
 }
 
 /// HTTPHeader represents an HTTP Header name and value as defined by RFC 7230.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesFiltersResponseHeaderModifierAdd {
+pub struct HTTPRouteRulesFiltersResponseHeaderModifierAdd {
     /// Name is the name of the HTTP Header to be matched. Name matching MUST be
     /// case insensitive. (See https://tools.ietf.org/html/rfc7230#section-3.2).
     ///
@@ -1378,7 +1758,7 @@ pub struct GRPCRouteRulesFiltersResponseHeaderModifierAdd {
 
 /// HTTPHeader represents an HTTP Header name and value as defined by RFC 7230.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesFiltersResponseHeaderModifierSet {
+pub struct HTTPRouteRulesFiltersResponseHeaderModifierSet {
     /// Name is the name of the HTTP Header to be matched. Name matching MUST be
     /// case insensitive. (See https://tools.ietf.org/html/rfc7230#section-3.2).
     ///
@@ -1392,230 +1772,358 @@ pub struct GRPCRouteRulesFiltersResponseHeaderModifierSet {
     pub value: String,
 }
 
-/// GRPCRouteFilter defines processing steps that must be completed during the
-/// request or response lifecycle. GRPCRouteFilters are meant as an extension
+/// HTTPRouteFilter defines processing steps that must be completed during the
+/// request or response lifecycle. HTTPRouteFilters are meant as an extension
 /// point to express processing that may be done in Gateway implementations. Some
 /// examples include request or response modification, implementing
 /// authentication strategies, rate-limiting, and traffic shaping. API
 /// guarantee/conformance is defined based on the type of the filter.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, PartialEq)]
-pub enum GRPCRouteRulesFiltersType {
-    ResponseHeaderModifier,
+pub enum HTTPRouteRulesFiltersType {
     RequestHeaderModifier,
+    ResponseHeaderModifier,
     RequestMirror,
+    RequestRedirect,
+    #[serde(rename = "URLRewrite")]
+    UrlRewrite,
     ExtensionRef,
 }
 
-/// GRPCRouteMatch defines the predicate used to match requests to a given
+/// URLRewrite defines a schema for a filter that modifies a request during forwarding.
+///
+/// Support: Extended
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
+pub struct HTTPRouteRulesFiltersUrlRewrite {
+    /// Hostname is the value to be used to replace the Host header value during
+    /// forwarding.
+    ///
+    /// Support: Extended
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub hostname: Option<String>,
+    /// Path defines a path rewrite.
+    ///
+    /// Support: Extended
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub path: Option<HTTPRouteRulesFiltersUrlRewritePath>,
+}
+
+/// Path defines a path rewrite.
+///
+/// Support: Extended
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
+pub struct HTTPRouteRulesFiltersUrlRewritePath {
+    /// ReplaceFullPath specifies the value with which to replace the full path
+    /// of a request during a rewrite or redirect.
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "replaceFullPath"
+    )]
+    pub replace_full_path: Option<String>,
+    /// ReplacePrefixMatch specifies the value with which to replace the prefix
+    /// match of a request during a rewrite or redirect. For example, a request
+    /// to "/foo/bar" with a prefix match of "/foo" and a ReplacePrefixMatch
+    /// of "/xyz" would be modified to "/xyz/bar".
+    ///
+    /// Note that this matches the behavior of the PathPrefix match type. This
+    /// matches full path elements. A path element refers to the list of labels
+    /// in the path split by the `/` separator. When specified, a trailing `/` is
+    /// ignored. For example, the paths `/abc`, `/abc/`, and `/abc/def` would all
+    /// match the prefix `/abc`, but the path `/abcd` would not.
+    ///
+    /// ReplacePrefixMatch is only compatible with a `PathPrefix` HTTPRouteMatch.
+    /// Using any other HTTPRouteMatch type on the same HTTPRouteRule will result in
+    /// the implementation setting the Accepted Condition for the Route to `status: False`.
+    ///
+    /// Request Path | Prefix Match | Replace Prefix | Modified Path
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "replacePrefixMatch"
+    )]
+    pub replace_prefix_match: Option<String>,
+    /// Type defines the type of path modifier. Additional types may be
+    /// added in a future release of the API.
+    ///
+    /// Note that values may be added to this enum, implementations
+    /// must ensure that unknown values will not cause a crash.
+    ///
+    /// Unknown values here must result in the implementation setting the
+    /// Accepted Condition for the Route to `status: False`, with a
+    /// Reason of `UnsupportedValue`.
+    #[serde(rename = "type")]
+    pub r#type: HTTPRouteRulesFiltersUrlRewritePathType,
+}
+
+/// Path defines a path rewrite.
+///
+/// Support: Extended
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, PartialEq)]
+pub enum HTTPRouteRulesFiltersUrlRewritePathType {
+    ReplaceFullPath,
+    ReplacePrefixMatch,
+}
+
+/// HTTPRouteMatch defines the predicate used to match requests to a given
 /// action. Multiple match types are ANDed together, i.e. the match will
 /// evaluate to true only if all conditions are satisfied.
 ///
-/// For example, the match below will match a gRPC request only if its service
-/// is `foo` AND it contains the `version: v1` header:
+/// For example, the match below will match a HTTP request only if its path
+/// starts with `/foo` AND it contains the `version: v1` header:
 ///
-/// ```
-/// matches:
-///   - method:
-///     type: Exact
-///     service: "foo"
-///     headers:
-///   - name: "version"
-///     value "v1"
+/// ```text
+/// match:
+///
+/// 	path:
+/// 	  value: "/foo"
+/// 	headers:
+/// 	- name: "version"
+/// 	  value "v1"
 ///
 /// ```
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesMatches {
-    /// Headers specifies gRPC request header matchers. Multiple match values are
-    /// ANDed together, meaning, a request MUST match all the specified headers
+pub struct HTTPRouteRulesMatches {
+    /// Headers specifies HTTP request header matchers. Multiple match values are
+    /// ANDed together, meaning, a request must match all the specified headers
     /// to select the route.
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub headers: Option<Vec<GRPCRouteRulesMatchesHeaders>>,
-    /// Method specifies a gRPC request service/method matcher. If this field is
-    /// not specified, all services and methods will match.
+    pub headers: Option<Vec<HTTPRouteRulesMatchesHeaders>>,
+    /// Method specifies HTTP method matcher.
+    /// When specified, this route will be matched only if the request has the
+    /// specified method.
+    ///
+    /// Support: Extended
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub method: Option<GRPCRouteRulesMatchesMethod>,
+    pub method: Option<HTTPRouteRulesMatchesMethod>,
+    /// Path specifies a HTTP request path matcher. If this field is not
+    /// specified, a default prefix match on the "/" path is provided.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub path: Option<HTTPRouteRulesMatchesPath>,
+    /// QueryParams specifies HTTP query parameter matchers. Multiple match
+    /// values are ANDed together, meaning, a request must match all the
+    /// specified query parameters to select the route.
+    ///
+    /// Support: Extended
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "queryParams"
+    )]
+    pub query_params: Option<Vec<HTTPRouteRulesMatchesQueryParams>>,
 }
 
-/// GRPCHeaderMatch describes how to select a gRPC route by matching gRPC request
+/// HTTPHeaderMatch describes how to select a HTTP route by matching HTTP request
 /// headers.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesMatchesHeaders {
-    /// Name is the name of the gRPC Header to be matched.
+pub struct HTTPRouteRulesMatchesHeaders {
+    /// Name is the name of the HTTP Header to be matched. Name matching MUST be
+    /// case insensitive. (See https://tools.ietf.org/html/rfc7230#section-3.2).
     ///
     /// If multiple entries specify equivalent header names, only the first
     /// entry with an equivalent name MUST be considered for a match. Subsequent
     /// entries with an equivalent header name MUST be ignored. Due to the
     /// case-insensitivity of header names, "foo" and "Foo" are considered
     /// equivalent.
+    ///
+    /// When a header is repeated in an HTTP request, it is
+    /// implementation-specific behavior as to how this is represented.
+    /// Generally, proxies should follow the guidance from the RFC:
+    /// https://www.rfc-editor.org/rfc/rfc7230.html#section-3.2.2 regarding
+    /// processing a repeated header, with special handling for "Set-Cookie".
     pub name: String,
     /// Type specifies how to match against the value of the header.
+    ///
+    /// Support: Core (Exact)
+    ///
+    /// Support: Implementation-specific (RegularExpression)
+    ///
+    /// Since RegularExpression HeaderMatchType has implementation-specific
+    /// conformance, implementations can support POSIX, PCRE or any other dialects
+    /// of regular expressions. Please read the implementation's documentation to
+    /// determine the supported dialect.
     #[serde(default, skip_serializing_if = "Option::is_none", rename = "type")]
-    pub r#type: Option<GRPCRouteRulesMatchesHeadersType>,
-    /// Value is the value of the gRPC Header to be matched.
+    pub r#type: Option<HTTPRouteRulesMatchesHeadersType>,
+    /// Value is the value of HTTP Header to be matched.
     pub value: String,
 }
 
-/// GRPCHeaderMatch describes how to select a gRPC route by matching gRPC request
+/// HTTPHeaderMatch describes how to select a HTTP route by matching HTTP request
 /// headers.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, PartialEq)]
-pub enum GRPCRouteRulesMatchesHeadersType {
+pub enum HTTPRouteRulesMatchesHeadersType {
     Exact,
     RegularExpression,
 }
 
-/// Method specifies a gRPC request service/method matcher. If this field is
-/// not specified, all services and methods will match.
+/// HTTPRouteMatch defines the predicate used to match requests to a given
+/// action. Multiple match types are ANDed together, i.e. the match will
+/// evaluate to true only if all conditions are satisfied.
+///
+/// For example, the match below will match a HTTP request only if its path
+/// starts with `/foo` AND it contains the `version: v1` header:
+///
+/// ```text
+/// match:
+///
+/// 	path:
+/// 	  value: "/foo"
+/// 	headers:
+/// 	- name: "version"
+/// 	  value "v1"
+///
+/// ```
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, PartialEq)]
+pub enum HTTPRouteRulesMatchesMethod {
+    #[serde(rename = "GET")]
+    Get,
+    #[serde(rename = "HEAD")]
+    Head,
+    #[serde(rename = "POST")]
+    Post,
+    #[serde(rename = "PUT")]
+    Put,
+    #[serde(rename = "DELETE")]
+    Delete,
+    #[serde(rename = "CONNECT")]
+    Connect,
+    #[serde(rename = "OPTIONS")]
+    Options,
+    #[serde(rename = "TRACE")]
+    Trace,
+    #[serde(rename = "PATCH")]
+    Patch,
+}
+
+/// Path specifies a HTTP request path matcher. If this field is not
+/// specified, a default prefix match on the "/" path is provided.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesMatchesMethod {
-    /// Value of the method to match against. If left empty or omitted, will
-    /// match all services.
+pub struct HTTPRouteRulesMatchesPath {
+    /// Type specifies how to match against the path Value.
     ///
-    /// At least one of Service and Method MUST be a non-empty string.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub method: Option<String>,
-    /// Value of the service to match against. If left empty or omitted, will
-    /// match any service.
-    ///
-    /// At least one of Service and Method MUST be a non-empty string.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub service: Option<String>,
-    /// Type specifies how to match against the service and/or method.
-    /// Support: Core (Exact with service and method specified)
-    ///
-    /// Support: Implementation-specific (Exact with method specified but no service specified)
+    /// Support: Core (Exact, PathPrefix)
     ///
     /// Support: Implementation-specific (RegularExpression)
     #[serde(default, skip_serializing_if = "Option::is_none", rename = "type")]
-    pub r#type: Option<GRPCRouteRulesMatchesMethodType>,
+    pub r#type: Option<HTTPRouteRulesMatchesPathType>,
+    /// Value of the HTTP path to match against.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub value: Option<String>,
 }
 
-/// Method specifies a gRPC request service/method matcher. If this field is
-/// not specified, all services and methods will match.
+/// Path specifies a HTTP request path matcher. If this field is not
+/// specified, a default prefix match on the "/" path is provided.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, PartialEq)]
-pub enum GRPCRouteRulesMatchesMethodType {
+pub enum HTTPRouteRulesMatchesPathType {
+    Exact,
+    PathPrefix,
+    RegularExpression,
+}
+
+/// HTTPQueryParamMatch describes how to select a HTTP route by matching HTTP
+/// query parameters.
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
+pub struct HTTPRouteRulesMatchesQueryParams {
+    /// Name is the name of the HTTP query param to be matched. This must be an
+    /// exact string match. (See
+    /// https://tools.ietf.org/html/rfc7230#section-2.7.3).
+    ///
+    /// If multiple entries specify equivalent query param names, only the first
+    /// entry with an equivalent name MUST be considered for a match. Subsequent
+    /// entries with an equivalent query param name MUST be ignored.
+    ///
+    /// If a query param is repeated in an HTTP request, the behavior is
+    /// purposely left undefined, since different data planes have different
+    /// capabilities. However, it is *recommended* that implementations should
+    /// match against the first value of the param if the data plane supports it,
+    /// as this behavior is expected in other load balancing contexts outside of
+    /// the Gateway API.
+    ///
+    /// Users SHOULD NOT route traffic based on repeated query params to guard
+    /// themselves against potential differences in the implementations.
+    pub name: String,
+    /// Type specifies how to match against the value of the query parameter.
+    ///
+    /// Support: Extended (Exact)
+    ///
+    /// Support: Implementation-specific (RegularExpression)
+    ///
+    /// Since RegularExpression QueryParamMatchType has Implementation-specific
+    /// conformance, implementations can support POSIX, PCRE or any other
+    /// dialects of regular expressions. Please read the implementation's
+    /// documentation to determine the supported dialect.
+    #[serde(default, skip_serializing_if = "Option::is_none", rename = "type")]
+    pub r#type: Option<HTTPRouteRulesMatchesQueryParamsType>,
+    /// Value is the value of HTTP query param to be matched.
+    pub value: String,
+}
+
+/// HTTPQueryParamMatch describes how to select a HTTP route by matching HTTP
+/// query parameters.
+#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, PartialEq)]
+pub enum HTTPRouteRulesMatchesQueryParamsType {
     Exact,
     RegularExpression,
 }
 
-/// SessionPersistence defines and configures session persistence
-/// for the route rule.
+/// Timeouts defines the timeouts that can be configured for an HTTP request.
 ///
 /// Support: Extended
-///
-///
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesSessionPersistence {
-    /// AbsoluteTimeout defines the absolute timeout of the persistent
-    /// session. Once the AbsoluteTimeout duration has elapsed, the
-    /// session becomes invalid.
+pub struct HTTPRouteRulesTimeouts {
+    /// BackendRequest specifies a timeout for an individual request from the gateway
+    /// to a backend. This covers the time from when the request first starts being
+    /// sent from the gateway to when the full response has been received from the backend.
+    ///
+    /// Setting a timeout to the zero duration (e.g. "0s") SHOULD disable the timeout
+    /// completely. Implementations that cannot completely disable the timeout MUST
+    /// instead interpret the zero duration as the longest possible value to which
+    /// the timeout can be set.
+    ///
+    /// An entire client HTTP transaction with a gateway, covered by the Request timeout,
+    /// may result in more than one call from the gateway to the destination backend,
+    /// for example, if automatic retries are supported.
+    ///
+    /// The value of BackendRequest must be a Gateway API Duration string as defined by
+    /// GEP-2257.  When this field is unspecified, its behavior is implementation-specific;
+    /// when specified, the value of BackendRequest must be no more than the value of the
+    /// Request timeout (since the Request timeout encompasses the BackendRequest timeout).
     ///
     /// Support: Extended
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
-        rename = "absoluteTimeout"
+        rename = "backendRequest"
     )]
-    pub absolute_timeout: Option<String>,
-    /// CookieConfig provides configuration settings that are specific
-    /// to cookie-based session persistence.
+    pub backend_request: Option<String>,
+    /// Request specifies the maximum duration for a gateway to respond to an HTTP request.
+    /// If the gateway has not been able to respond before this deadline is met, the gateway
+    /// MUST return a timeout error.
     ///
-    /// Support: Core
-    #[serde(
-        default,
-        skip_serializing_if = "Option::is_none",
-        rename = "cookieConfig"
-    )]
-    pub cookie_config: Option<GRPCRouteRulesSessionPersistenceCookieConfig>,
-    /// IdleTimeout defines the idle timeout of the persistent session.
-    /// Once the session has been idle for more than the specified
-    /// IdleTimeout duration, the session becomes invalid.
+    /// For example, setting the `rules.timeouts.request` field to the value `10s` in an
+    /// `HTTPRoute` will cause a timeout if a client request is taking longer than 10 seconds
+    /// to complete.
+    ///
+    /// Setting a timeout to the zero duration (e.g. "0s") SHOULD disable the timeout
+    /// completely. Implementations that cannot completely disable the timeout MUST
+    /// instead interpret the zero duration as the longest possible value to which
+    /// the timeout can be set.
+    ///
+    /// This timeout is intended to cover as close to the whole request-response transaction
+    /// as possible although an implementation MAY choose to start the timeout after the entire
+    /// request stream has been received instead of immediately after the transaction is
+    /// initiated by the client.
+    ///
+    /// The value of Request is a Gateway API Duration string as defined by GEP-2257. When this
+    /// field is unspecified, request timeout behavior is implementation-specific.
     ///
     /// Support: Extended
-    #[serde(
-        default,
-        skip_serializing_if = "Option::is_none",
-        rename = "idleTimeout"
-    )]
-    pub idle_timeout: Option<String>,
-    /// SessionName defines the name of the persistent session token
-    /// which may be reflected in the cookie or the header. Users
-    /// should avoid reusing session names to prevent unintended
-    /// consequences, such as rejection or unpredictable behavior.
-    ///
-    /// Support: Implementation-specific
-    #[serde(
-        default,
-        skip_serializing_if = "Option::is_none",
-        rename = "sessionName"
-    )]
-    pub session_name: Option<String>,
-    /// Type defines the type of session persistence such as through
-    /// the use a header or cookie. Defaults to cookie based session
-    /// persistence.
-    ///
-    /// Support: Core for "Cookie" type
-    ///
-    /// Support: Extended for "Header" type
-    #[serde(default, skip_serializing_if = "Option::is_none", rename = "type")]
-    pub r#type: Option<GRPCRouteRulesSessionPersistenceType>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub request: Option<String>,
 }
 
-/// CookieConfig provides configuration settings that are specific
-/// to cookie-based session persistence.
-///
-/// Support: Core
+/// Status defines the current state of HTTPRoute.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteRulesSessionPersistenceCookieConfig {
-    /// LifetimeType specifies whether the cookie has a permanent or
-    /// session-based lifetime. A permanent cookie persists until its
-    /// specified expiry time, defined by the Expires or Max-Age cookie
-    /// attributes, while a session cookie is deleted when the current
-    /// session ends.
-    ///
-    /// When set to "Permanent", AbsoluteTimeout indicates the
-    /// cookie's lifetime via the Expires or Max-Age cookie attributes
-    /// and is required.
-    ///
-    /// When set to "Session", AbsoluteTimeout indicates the
-    /// absolute lifetime of the cookie tracked by the gateway and
-    /// is optional.
-    ///
-    /// Support: Core for "Session" type
-    ///
-    /// Support: Extended for "Permanent" type
-    #[serde(
-        default,
-        skip_serializing_if = "Option::is_none",
-        rename = "lifetimeType"
-    )]
-    pub lifetime_type: Option<GRPCRouteRulesSessionPersistenceCookieConfigLifetimeType>,
-}
-
-/// CookieConfig provides configuration settings that are specific
-/// to cookie-based session persistence.
-///
-/// Support: Core
-#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, PartialEq)]
-pub enum GRPCRouteRulesSessionPersistenceCookieConfigLifetimeType {
-    Permanent,
-    Session,
-}
-
-/// SessionPersistence defines and configures session persistence
-/// for the route rule.
-///
-/// Support: Extended
-///
-///
-#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, PartialEq)]
-pub enum GRPCRouteRulesSessionPersistenceType {
-    Cookie,
-    Header,
-}
-
-/// Status defines the current state of GRPCRoute.
-#[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteStatus {
+pub struct HTTPRouteStatus {
     /// Parents is a list of parent resources (usually Gateways) that are
     /// associated with the route, and the status of the route with respect to
     /// each parent. When this route attaches to a parent, the controller that
@@ -1630,13 +2138,13 @@ pub struct GRPCRouteStatus {
     ///
     /// A maximum of 32 Gateways will be represented in this list. An empty list
     /// means the route has not been attached to any Gateway.
-    pub parents: Vec<GRPCRouteStatusParents>,
+    pub parents: Vec<HTTPRouteStatusParents>,
 }
 
 /// RouteParentStatus describes the status of a route with respect to an
 /// associated Parent.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteStatusParents {
+pub struct HTTPRouteStatusParents {
     /// Conditions describes the status of the route with respect to the Gateway.
     /// Note that the route's availability is also subject to the Gateway's own
     /// status conditions and listener status.
@@ -1676,13 +2184,13 @@ pub struct GRPCRouteStatusParents {
     /// ParentRef corresponds with a ParentRef in the spec that this
     /// RouteParentStatus struct describes the status of.
     #[serde(rename = "parentRef")]
-    pub parent_ref: GRPCRouteStatusParentsParentRef,
+    pub parent_ref: HTTPRouteStatusParentsParentRef,
 }
 
 /// ParentRef corresponds with a ParentRef in the spec that this
 /// RouteParentStatus struct describes the status of.
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema, Default, PartialEq)]
-pub struct GRPCRouteStatusParentsParentRef {
+pub struct HTTPRouteStatusParentsParentRef {
     /// Group is the group of the referent.
     /// When unspecified, "gateway.networking.k8s.io" is inferred.
     /// To set the core API group (such as for a "Service" kind referent),
@@ -1715,16 +2223,6 @@ pub struct GRPCRouteStatusParentsParentRef {
     /// generic way to enable any other kind of cross-namespace reference.
     ///
     ///
-    /// ParentRefs from a Route to a Service in the same namespace are "producer"
-    /// routes, which apply default routing rules to inbound connections from
-    /// any namespace to the Service.
-    ///
-    /// ParentRefs from a Route to a Service in a different namespace are
-    /// "consumer" routes, and these routing rules are only applied to outbound
-    /// connections originating from the same namespace as the Route, for which
-    /// the intended destination of the connections are a Service targeted as a
-    /// ParentRef of the Route.
-    ///
     ///
     /// Support: Core
     #[serde(default, skip_serializing_if = "Option::is_none")]
@@ -1740,10 +2238,6 @@ pub struct GRPCRouteStatusParentsParentRef {
     /// and SectionName are specified, the name and port of the selected listener
     /// must match both specified values.
     ///
-    ///
-    /// When the parent resource is a Service, this targets a specific port in the
-    /// Service spec. When both Port (experimental) and SectionName are specified,
-    /// the name and port of the selected port must match both specified values.
     ///
     ///
     /// Implementations MAY choose to support other parent resources.
